@@ -3,34 +3,35 @@
 	<PropertyGroup>
 		<GenerateSatelliteAssembliesForCore>true</GenerateSatelliteAssembliesForCore>
 		<OutputType>WinExe</OutputType>
-		<TargetFramework>net6.0-windows</TargetFramework>
+		<TargetFramework>net8.0-windows</TargetFramework>
 		<Nullable>enable</Nullable>
 		<UseWPF>true</UseWPF>  
 		<ApplicationManifest>app.manifest</ApplicationManifest>
 		<ImplicitUsings>enable</ImplicitUsings>
 		<ApplicationIcon>v2rayN.ico</ApplicationIcon>
-		<Copyright>Copyright © 2017-2023 (GPLv3)</Copyright>
-		<FileVersion>6.29</FileVersion>
+		<Copyright>Copyright © 2017-2024 (GPLv3)</Copyright>
+		<FileVersion>6.33</FileVersion>
 	</PropertyGroup>
-
+	
 	<ItemGroup>
 		<PackageReference Include="Downloader" Version="3.0.6" />		
 		<PackageReference Include="MaterialDesignThemes" Version="4.9.0" />
-		<PackageReference Include="H.NotifyIcon.Wpf" Version="2.0.118" />
+		<PackageReference Include="H.NotifyIcon.Wpf" Version="2.0.124" />
 		<PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
 		<PackageReference Include="QRCoder.Xaml" Version="1.4.3" />
 		<PackageReference Include="sqlite-net-pcl" Version="1.8.116" />
 		<PackageReference Include="TaskScheduler" Version="2.10.1" /> 
 		<PackageReference Include="ZXing.Net.Bindings.Windows.Compatibility" Version="0.16.12" />
-		<PackageReference Include="ReactiveUI.Fody" Version="18.4.1" />
-		<PackageReference Include="ReactiveUI.Validation" Version="3.0.22" />
-		<PackageReference Include="ReactiveUI.WPF" Version="18.4.1" />
-		<PackageReference Include="Splat.NLog" Version="14.7.1" />
+		<PackageReference Include="ReactiveUI.Fody" Version="19.5.39" />
+		<PackageReference Include="ReactiveUI.Validation" Version="3.1.7" />
+		<PackageReference Include="ReactiveUI.WPF" Version="19.5.39" />
+		<PackageReference Include="Splat.NLog" Version="14.8.12" />
 		<PackageReference Include="System.Reactive" Version="6.0.0" />
 	</ItemGroup>
 	
 	<ItemGroup>       
         <AdditionalFiles Include="app.manifest" />       
+        <EmbeddedResource Include="Sample\SingboxSampleOutbound" />       
         <EmbeddedResource Include="Sample\SingboxSampleClientConfig">
           <CopyToOutputDirectory>Never</CopyToOutputDirectory>
         </EmbeddedResource>
@@ -63,8 +64,9 @@
         </EmbeddedResource>  
         <EmbeddedResource Include="Sample\SampleInbound">
           <CopyToOutputDirectory>Never</CopyToOutputDirectory>
-        </EmbeddedResource>  
-        <EmbeddedResource Include="Sample\tun_singbox_rules">
+        </EmbeddedResource>
+		<EmbeddedResource Include="Sample\SampleOutbound" />
+		<EmbeddedResource Include="Sample\tun_singbox_rules">
           <CopyToOutputDirectory>Never</CopyToOutputDirectory>
         </EmbeddedResource>  
         <EmbeddedResource Include="Sample\tun_singbox_inbound">
