@@ -2,15 +2,19 @@
 using QRCoder.Xaml;
 using System.Windows.Media;
 
-namespace v2rayN.Handler
+namespace v2rayN
 {
     /// <summary>
     /// 含有QR码的描述类和包装编码和渲染
     /// </summary>
     public class QRCodeHelper
     {
-        public static DrawingImage? GetQRCode(string strContent)
+        public static DrawingImage? GetQRCode(string? strContent)
         {
+            if (strContent is null)
+            {
+                return null;
+            }
             try
             {
                 QRCodeGenerator qrGenerator = new();
