@@ -1,9 +1,7 @@
 ﻿using System.Data;
 using System.IO;
 using System.Text.RegularExpressions;
-using v2rayN.Base;
 using v2rayN.Mode;
-using v2rayN.Tool;
 
 namespace v2rayN.Handler
 {
@@ -22,20 +20,20 @@ namespace v2rayN.Handler
         /// </summary>
         /// <param name="config"></param>
         /// <returns></returns>
-        public static int LoadConfig(ref Config config)
+        public static int LoadConfig(ref Config? config)
         {
             //载入配置文件
-            string result = Utils.LoadResource(Utils.GetConfigPath(configRes));
+            var result = Utils.LoadResource(Utils.GetConfigPath(configRes));
             if (!Utils.IsNullOrEmpty(result))
             {
                 //转成Json
-                config = Utils.FromJson<Config>(result);
+                config = JsonUtils.Deserialize<Config>(result);
             }
             else
             {
                 if (File.Exists(Utils.GetConfigPath(configRes)))
                 {
-                    Utils.SaveLog("LoadConfig Exception");
+                    Logging.SaveLog("LoadConfig Exception");
                     return -1;
                 }
             }
@@ -96,7 +94,7 @@ namespace v2rayN.Handler
             //路由规则
             if (Utils.IsNullOrEmpty(config.routingBasicItem.domainStrategy))
             {
-                config.routingBasicItem.domainStrategy = Global.domainStrategys[0];//"IPIfNonMatch";
+                config.routingBasicItem.domainStrategy = Global.DomainStrategys[0];//"IPIfNonMatch";
             }
             //if (Utils.IsNullOrEmpty(config.domainMatcher))
             //{
@@ -184,15 +182,24 @@ namespace v2rayN.Handler
                 config.speedTestItem.speedPingTestUrl = Global.SpeedPingTestUrl;
             }
 
-            if (config.mux4Sbox == null)
+            if (config.mux4SboxItem == null)
             {
-                config.mux4Sbox = new()
+                config.mux4SboxItem = new()
                 {
                     protocol = Global.SingboxMuxs[0],
                     max_connections = 4,
                     min_streams = 4,
                     max_streams = 0,
                     padding = true
+                };
+            }
+
+            if (config.hysteriaItem == null)
+            {
+                config.hysteriaItem = new()
+                {
+                    up_mbps = 100,
+                    down_mbps = 100
                 };
             }
 
@@ -205,7 +212,7 @@ namespace v2rayN.Handler
         /// </summary>
         /// <param name="config"></param>
         /// <returns></returns>
-        public static int SaveConfig(ref Config config, bool reload = true)
+        public static int SaveConfig(Config config, bool reload = true)
         {
             ToJsonFile(config);
 
@@ -225,7 +232,7 @@ namespace v2rayN.Handler
                     //save temp file
                     var resPath = Utils.GetConfigPath(configRes);
                     var tempPath = $"{resPath}_temp";
-                    if (Utils.ToJsonFile(config, tempPath) != 0)
+                    if (JsonUtils.ToFile(config, tempPath) != 0)
                     {
                         return;
                     }
@@ -239,26 +246,26 @@ namespace v2rayN.Handler
                 }
                 catch (Exception ex)
                 {
-                    Utils.SaveLog("ToJsonFile", ex);
+                    Logging.SaveLog("ToJsonFile", ex);
                 }
             }
         }
 
-        public static int ImportOldGuiConfig(ref Config config, string fileName)
+        public static int ImportOldGuiConfig(Config config, string fileName)
         {
-            string result = Utils.LoadResource(fileName);
+            var result = Utils.LoadResource(fileName);
             if (Utils.IsNullOrEmpty(result))
             {
                 return -1;
             }
 
-            var configOld = Utils.FromJson<ConfigOld>(result);
+            var configOld = JsonUtils.Deserialize<ConfigOld>(result);
             if (configOld == null)
             {
                 return -1;
             }
 
-            var subItem = Utils.FromJson<List<SubItem>>(Utils.ToJson(configOld.subItem));
+            var subItem = JsonUtils.Deserialize<List<SubItem>>(JsonUtils.Serialize(configOld.subItem));
             foreach (var it in subItem)
             {
                 if (Utils.IsNullOrEmpty(it.id))
@@ -268,7 +275,7 @@ namespace v2rayN.Handler
                 SqliteHelper.Instance.Replace(it);
             }
 
-            var profileItems = Utils.FromJson<List<ProfileItem>>(Utils.ToJson(configOld.vmess));
+            var profileItems = JsonUtils.Deserialize<List<ProfileItem>>(JsonUtils.Serialize(configOld.vmess));
             foreach (var it in profileItems)
             {
                 if (Utils.IsNullOrEmpty(it.indexId))
@@ -284,13 +291,13 @@ namespace v2rayN.Handler
                 {
                     continue;
                 }
-                var routing = Utils.FromJson<RoutingItem>(Utils.ToJson(it));
+                var routing = JsonUtils.Deserialize<RoutingItem>(JsonUtils.Serialize(it));
                 foreach (var it2 in it.rules)
                 {
                     it2.id = Utils.GetGUID(false);
                 }
                 routing.ruleNum = it.rules.Count;
-                routing.ruleSet = Utils.ToJson(it.rules, false);
+                routing.ruleSet = JsonUtils.Serialize(it.rules, false);
 
                 if (Utils.IsNullOrEmpty(routing.id))
                 {
@@ -299,7 +306,7 @@ namespace v2rayN.Handler
                 SqliteHelper.Instance.Replace(routing);
             }
 
-            config = Utils.FromJson<Config>(Utils.ToJson(configOld));
+            config = JsonUtils.Deserialize<Config>(JsonUtils.Serialize(configOld));
 
             if (config.coreBasicItem == null)
             {
@@ -334,9 +341,9 @@ namespace v2rayN.Handler
                 };
             }
 
-            GetDefaultServer(ref config);
-            GetDefaultRouting(ref config);
-            SaveConfig(ref config);
+            GetDefaultServer(config);
+            GetDefaultRouting(config);
+            SaveConfig(config);
             LoadConfig(ref config);
 
             return 0;
@@ -347,12 +354,12 @@ namespace v2rayN.Handler
         #region Server
 
         /// <summary>
-        /// 添加服务器或编辑
+        /// Add or edit server
         /// </summary>
         /// <param name="config"></param>
         /// <param name="profileItem"></param>
         /// <returns></returns>
-        public static int AddServer(ref Config config, ProfileItem profileItem, bool toFile = true)
+        public static int AddServer(Config config, ProfileItem profileItem, bool toFile = true)
         {
             profileItem.configType = EConfigType.VMess;
 
@@ -365,7 +372,7 @@ namespace v2rayN.Handler
             profileItem.path = profileItem.path.TrimEx();
             profileItem.streamSecurity = profileItem.streamSecurity.TrimEx();
 
-            if (!Global.vmessSecuritys.Contains(profileItem.security))
+            if (!Global.VmessSecuritys.Contains(profileItem.security))
             {
                 return -1;
             }
@@ -374,7 +381,7 @@ namespace v2rayN.Handler
                 return -1;
             }
 
-            AddServerCommon(ref config, profileItem, toFile);
+            AddServerCommon(config, profileItem, toFile);
 
             return 0;
         }
@@ -394,7 +401,7 @@ namespace v2rayN.Handler
             }
 
             SqliteHelper.Instance.UpdateAll(indexs);
-            RemoveServerViaSubid(ref config, subid, false);
+            RemoveServerViaSubid(config, subid, false);
 
             return 0;
         }
@@ -405,7 +412,7 @@ namespace v2rayN.Handler
         /// <param name="config"></param>
         /// <param name="index"></param>
         /// <returns></returns>
-        public static int CopyServer(ref Config config, List<ProfileItem> indexs)
+        public static int CopyServer(Config config, List<ProfileItem> indexs)
         {
             foreach (var it in indexs)
             {
@@ -415,20 +422,20 @@ namespace v2rayN.Handler
                     continue;
                 }
 
-                ProfileItem profileItem = Utils.DeepCopy(item);
+                ProfileItem profileItem = JsonUtils.DeepCopy(item);
                 profileItem.indexId = string.Empty;
                 profileItem.remarks = $"{item.remarks}-clone";
 
                 if (profileItem.configType == EConfigType.Custom)
                 {
                     profileItem.address = Utils.GetConfigPath(profileItem.address);
-                    if (AddCustomServer(ref config, profileItem, false) == 0)
+                    if (AddCustomServer(config, profileItem, false) == 0)
                     {
                     }
                 }
                 else
                 {
-                    AddServerCommon(ref config, profileItem, true);
+                    AddServerCommon(config, profileItem, true);
                 }
             }
 
@@ -441,7 +448,7 @@ namespace v2rayN.Handler
         /// <param name="config"></param>
         /// <param name="item"></param>
         /// <returns></returns>
-        public static int SetDefaultServerIndex(ref Config config, string? indexId)
+        public static int SetDefaultServerIndex(Config config, string? indexId)
         {
             if (Utils.IsNullOrEmpty(indexId))
             {
@@ -467,18 +474,18 @@ namespace v2rayN.Handler
             }
             if (lstProfile.Count > 0)
             {
-                return SetDefaultServerIndex(ref config, lstProfile.Where(t => t.port > 0).FirstOrDefault()?.indexId);
+                return SetDefaultServerIndex(config, lstProfile.Where(t => t.port > 0).FirstOrDefault()?.indexId);
             }
-            return SetDefaultServerIndex(ref config, SqliteHelper.Instance.Table<ProfileItem>().Where(t => t.port > 0).Select(t => t.indexId).FirstOrDefault());
+            return SetDefaultServerIndex(config, SqliteHelper.Instance.Table<ProfileItem>().Where(t => t.port > 0).Select(t => t.indexId).FirstOrDefault());
         }
 
-        public static ProfileItem? GetDefaultServer(ref Config config)
+        public static ProfileItem? GetDefaultServer(Config config)
         {
             var item = LazyConfig.Instance.GetProfileItem(config.indexId);
             if (item is null)
             {
                 var item2 = SqliteHelper.Instance.Table<ProfileItem>().FirstOrDefault();
-                SetDefaultServerIndex(ref config, item2?.indexId);
+                SetDefaultServerIndex(config, item2?.indexId);
                 return item2;
             }
 
@@ -493,7 +500,7 @@ namespace v2rayN.Handler
         /// <param name="index"></param>
         /// <param name="eMove"></param>
         /// <returns></returns>
-        public static int MoveServer(ref Config config, ref List<ProfileItem> lstProfile, int index, EMove eMove, int pos = -1)
+        public static int MoveServer(Config config, ref List<ProfileItem> lstProfile, int index, EMove eMove, int pos = -1)
         {
             int count = lstProfile.Count;
             if (index < 0 || index > lstProfile.Count - 1)
@@ -565,7 +572,7 @@ namespace v2rayN.Handler
         /// <param name="config"></param>
         /// <param name="profileItem"></param>
         /// <returns></returns>
-        public static int AddCustomServer(ref Config config, ProfileItem profileItem, bool blDelete)
+        public static int AddCustomServer(Config config, ProfileItem profileItem, bool blDelete)
         {
             var fileName = profileItem.address;
             if (!File.Exists(fileName))
@@ -586,7 +593,7 @@ namespace v2rayN.Handler
             }
             catch (Exception ex)
             {
-                Utils.SaveLog(ex.Message, ex);
+                Logging.SaveLog(ex.Message, ex);
                 return -1;
             }
 
@@ -597,18 +604,18 @@ namespace v2rayN.Handler
                 profileItem.remarks = $"import custom@{DateTime.Now.ToString("yyyy/MM/dd HH:mm:ss")}";
             }
 
-            AddServerCommon(ref config, profileItem, true);
+            AddServerCommon(config, profileItem, true);
 
             return 0;
         }
 
         /// <summary>
-        /// 添加服务器或编辑
+        /// Add or edit server
         /// </summary>
         /// <param name="config"></param>
         /// <param name="profileItem"></param>
         /// <returns></returns>
-        public static int EditCustomServer(ref Config config, ProfileItem profileItem)
+        public static int EditCustomServer(Config config, ProfileItem profileItem)
         {
             if (SqliteHelper.Instance.Update(profileItem) > 0)
             {
@@ -623,12 +630,12 @@ namespace v2rayN.Handler
         }
 
         /// <summary>
-        /// 添加服务器或编辑
+        /// Add or edit server
         /// </summary>
         /// <param name="config"></param>
         /// <param name="profileItem"></param>
         /// <returns></returns>
-        public static int AddShadowsocksServer(ref Config config, ProfileItem profileItem, bool toFile = true)
+        public static int AddShadowsocksServer(Config config, ProfileItem profileItem, bool toFile = true)
         {
             profileItem.configType = EConfigType.Shadowsocks;
 
@@ -645,35 +652,35 @@ namespace v2rayN.Handler
                 return -1;
             }
 
-            AddServerCommon(ref config, profileItem, toFile);
+            AddServerCommon(config, profileItem, toFile);
 
             return 0;
         }
 
         /// <summary>
-        /// 添加服务器或编辑
+        /// Add or edit server
         /// </summary>
         /// <param name="config"></param>
         /// <param name="profileItem"></param>
         /// <returns></returns>
-        public static int AddSocksServer(ref Config config, ProfileItem profileItem, bool toFile = true)
+        public static int AddSocksServer(Config config, ProfileItem profileItem, bool toFile = true)
         {
             profileItem.configType = EConfigType.Socks;
 
             profileItem.address = profileItem.address.TrimEx();
 
-            AddServerCommon(ref config, profileItem, toFile);
+            AddServerCommon(config, profileItem, toFile);
 
             return 0;
         }
 
         /// <summary>
-        /// 添加服务器或编辑
+        /// Add or edit server
         /// </summary>
         /// <param name="config"></param>
         /// <param name="profileItem"></param>
         /// <returns></returns>
-        public static int AddTrojanServer(ref Config config, ProfileItem profileItem, bool toFile = true)
+        public static int AddTrojanServer(Config config, ProfileItem profileItem, bool toFile = true)
         {
             profileItem.configType = EConfigType.Trojan;
 
@@ -688,12 +695,113 @@ namespace v2rayN.Handler
                 return -1;
             }
 
-            AddServerCommon(ref config, profileItem, toFile);
+            AddServerCommon(config, profileItem, toFile);
 
             return 0;
         }
 
-        public static int SortServers(ref Config config, string subId, string colName, bool asc)
+        /// <summary>
+        /// Add or edit server
+        /// </summary>
+        /// <param name="config"></param>
+        /// <param name="profileItem"></param>
+        /// <returns></returns>
+        public static int AddHysteria2Server(Config config, ProfileItem profileItem, bool toFile = true)
+        {
+            profileItem.configType = EConfigType.Hysteria2;
+            profileItem.coreType = ECoreType.sing_box;
+
+            profileItem.address = profileItem.address.TrimEx();
+            profileItem.id = profileItem.id.TrimEx();
+            profileItem.path = profileItem.path.TrimEx();
+            profileItem.network = string.Empty;
+
+            if (Utils.IsNullOrEmpty(profileItem.streamSecurity))
+            {
+                profileItem.streamSecurity = Global.StreamSecurity;
+            }
+            if (profileItem.id.IsNullOrEmpty())
+            {
+                return -1;
+            }
+
+            AddServerCommon(config, profileItem, toFile);
+
+            return 0;
+        }
+
+        /// <summary>
+        /// Add or edit server
+        /// </summary>
+        /// <param name="config"></param>
+        /// <param name="profileItem"></param>
+        /// <returns></returns>
+        public static int AddTuicServer(Config config, ProfileItem profileItem, bool toFile = true)
+        {
+            profileItem.configType = EConfigType.Tuic;
+            profileItem.coreType = ECoreType.sing_box;
+
+            profileItem.address = profileItem.address.TrimEx();
+            profileItem.id = profileItem.id.TrimEx();
+            profileItem.security = profileItem.security.TrimEx();
+            profileItem.network = string.Empty;
+
+            if (!Global.TuicCongestionControls.Contains(profileItem.headerType))
+            {
+                profileItem.headerType = Global.TuicCongestionControls.FirstOrDefault()!;
+            }
+
+            if (Utils.IsNullOrEmpty(profileItem.streamSecurity))
+            {
+                profileItem.streamSecurity = Global.StreamSecurity;
+            }
+            if (Utils.IsNullOrEmpty(profileItem.alpn))
+            {
+                profileItem.alpn = "h3";
+            }
+            if (profileItem.id.IsNullOrEmpty())
+            {
+                return -1;
+            }
+
+            AddServerCommon(config, profileItem, toFile);
+
+            return 0;
+        }
+
+        /// <summary>
+        /// Add or edit server
+        /// </summary>
+        /// <param name="config"></param>
+        /// <param name="profileItem"></param>
+        /// <returns></returns>
+        public static int AddWireguardServer(Config config, ProfileItem profileItem, bool toFile = true)
+        {
+            profileItem.configType = EConfigType.Wireguard;
+            profileItem.coreType = ECoreType.sing_box;
+
+            profileItem.address = profileItem.address.TrimEx();
+            profileItem.id = profileItem.id.TrimEx();
+            profileItem.publicKey = profileItem.publicKey.TrimEx();
+            profileItem.path = profileItem.path.TrimEx();
+            profileItem.requestHost = profileItem.requestHost.TrimEx();
+            profileItem.network = string.Empty;
+            if (profileItem.shortId.IsNullOrEmpty())
+            {
+                profileItem.shortId = Global.TunMtus.FirstOrDefault();
+            }
+
+            if (profileItem.id.IsNullOrEmpty())
+            {
+                return -1;
+            }
+
+            AddServerCommon(config, profileItem, toFile);
+
+            return 0;
+        }
+
+        public static int SortServers(Config config, string subId, string colName, bool asc)
         {
             var lstModel = LazyConfig.Instance.ProfileItems(subId, "");
             if (lstModel.Count <= 0)
@@ -790,12 +898,12 @@ namespace v2rayN.Handler
         }
 
         /// <summary>
-        /// 添加服务器或编辑
+        /// Add or edit server
         /// </summary>
         /// <param name="config"></param>
         /// <param name="profileItem"></param>
         /// <returns></returns>
-        public static int AddVlessServer(ref Config config, ProfileItem profileItem, bool toFile = true)
+        public static int AddVlessServer(Config config, ProfileItem profileItem, bool toFile = true)
         {
             profileItem.configType = EConfigType.VLESS;
 
@@ -808,16 +916,16 @@ namespace v2rayN.Handler
             profileItem.path = profileItem.path.TrimEx();
             profileItem.streamSecurity = profileItem.streamSecurity.TrimEx();
 
-            if (!Global.flows.Contains(profileItem.flow))
+            if (!Global.Flows.Contains(profileItem.flow))
             {
-                profileItem.flow = Global.flows.First();
+                profileItem.flow = Global.Flows.First();
             }
             if (profileItem.id.IsNullOrEmpty())
             {
                 return -1;
             }
 
-            AddServerCommon(ref config, profileItem, toFile);
+            AddServerCommon(config, profileItem, toFile);
 
             return 0;
         }
@@ -846,7 +954,7 @@ namespace v2rayN.Handler
             return new Tuple<int, int>(lstProfile.Count, lstKeep.Count);
         }
 
-        public static int AddServerCommon(ref Config config, ProfileItem profileItem, bool toFile = true)
+        public static int AddServerCommon(Config config, ProfileItem profileItem, bool toFile = true)
         {
             profileItem.configVersion = 2;
 
@@ -856,13 +964,13 @@ namespace v2rayN.Handler
                 {
                     profileItem.allowInsecure = config.coreBasicItem.defAllowInsecure.ToString().ToLower();
                 }
-                if (Utils.IsNullOrEmpty(profileItem.fingerprint))
+                if (Utils.IsNullOrEmpty(profileItem.fingerprint) && profileItem.streamSecurity == Global.StreamSecurityReality)
                 {
                     profileItem.fingerprint = config.coreBasicItem.defFingerprint;
                 }
             }
 
-            if (!Utils.IsNullOrEmpty(profileItem.network) && !Global.networks.Contains(profileItem.network))
+            if (!Utils.IsNullOrEmpty(profileItem.network) && !Global.Networks.Contains(profileItem.network))
             {
                 profileItem.network = Global.DefaultNetwork;
             }
@@ -930,7 +1038,7 @@ namespace v2rayN.Handler
             }
             catch (Exception ex)
             {
-                Utils.SaveLog("Remove Item", ex);
+                Logging.SaveLog("Remove Item", ex);
             }
 
             return 0;
@@ -947,7 +1055,7 @@ namespace v2rayN.Handler
         /// <param name="clipboardData"></param>
         /// <param name="subid"></param>
         /// <returns>成功导入的数量</returns>
-        private static int AddBatchServers(ref Config config, string clipboardData, string subid, bool isSub, List<ProfileItem> lstOriSub)
+        private static int AddBatchServers(Config config, string clipboardData, string subid, bool isSub, List<ProfileItem> lstOriSub)
         {
             if (Utils.IsNullOrEmpty(clipboardData))
             {
@@ -958,7 +1066,7 @@ namespace v2rayN.Handler
             //remove sub items
             if (isSub && !Utils.IsNullOrEmpty(subid))
             {
-                RemoveServerViaSubid(ref config, subid, isSub);
+                RemoveServerViaSubid(config, subid, isSub);
                 subFilter = LazyConfig.Instance.GetSubItem(subid)?.filter ?? "";
             }
 
@@ -974,9 +1082,9 @@ namespace v2rayN.Handler
             foreach (string str in arrData)
             {
                 //maybe sub
-                if (!isSub && (str.StartsWith(Global.httpsProtocol) || str.StartsWith(Global.httpProtocol)))
+                if (!isSub && (str.StartsWith(Global.HttpsProtocol) || str.StartsWith(Global.HttpProtocol)))
                 {
-                    if (AddSubItem(ref config, str) == 0)
+                    if (AddSubItem(config, str) == 0)
                     {
                         countServers++;
                     }
@@ -1024,23 +1132,35 @@ namespace v2rayN.Handler
 
                 if (profileItem.configType == EConfigType.VMess)
                 {
-                    addStatus = AddServer(ref config, profileItem, false);
+                    addStatus = AddServer(config, profileItem, false);
                 }
                 else if (profileItem.configType == EConfigType.Shadowsocks)
                 {
-                    addStatus = AddShadowsocksServer(ref config, profileItem, false);
+                    addStatus = AddShadowsocksServer(config, profileItem, false);
                 }
                 else if (profileItem.configType == EConfigType.Socks)
                 {
-                    addStatus = AddSocksServer(ref config, profileItem, false);
+                    addStatus = AddSocksServer(config, profileItem, false);
                 }
                 else if (profileItem.configType == EConfigType.Trojan)
                 {
-                    addStatus = AddTrojanServer(ref config, profileItem, false);
+                    addStatus = AddTrojanServer(config, profileItem, false);
                 }
                 else if (profileItem.configType == EConfigType.VLESS)
                 {
-                    addStatus = AddVlessServer(ref config, profileItem, false);
+                    addStatus = AddVlessServer(config, profileItem, false);
+                }
+                else if (profileItem.configType == EConfigType.Hysteria2)
+                {
+                    addStatus = AddHysteria2Server(config, profileItem, false);
+                }
+                else if (profileItem.configType == EConfigType.Tuic)
+                {
+                    addStatus = AddTuicServer(config, profileItem, false);
+                }
+                else if (profileItem.configType == EConfigType.Wireguard)
+                {
+                    addStatus = AddWireguardServer(config, profileItem, false);
                 }
 
                 if (addStatus == 0)
@@ -1059,7 +1179,7 @@ namespace v2rayN.Handler
             return countServers;
         }
 
-        private static int AddBatchServers4Custom(ref Config config, string clipboardData, string subid, bool isSub, List<ProfileItem> lstOriSub)
+        private static int AddBatchServers4Custom(Config config, string clipboardData, string subid, bool isSub, List<ProfileItem> lstOriSub)
         {
             if (Utils.IsNullOrEmpty(clipboardData))
             {
@@ -1078,7 +1198,7 @@ namespace v2rayN.Handler
 
             ProfileItem profileItem = new();
             //Is v2ray configuration
-            V2rayConfig? v2rayConfig = Utils.FromJson<V2rayConfig>(clipboardData);
+            V2rayConfig? v2rayConfig = JsonUtils.Deserialize<V2rayConfig>(clipboardData);
             if (v2rayConfig?.inbounds?.Count > 0
                 && v2rayConfig.outbounds?.Count > 0)
             {
@@ -1095,7 +1215,7 @@ namespace v2rayN.Handler
                 var fileName = Utils.GetTempPath($"{Utils.GetGUID(false)}.yaml");
                 File.WriteAllText(fileName, clipboardData);
 
-                profileItem.coreType = ECoreType.clash;
+                profileItem.coreType = ECoreType.mihomo;
                 profileItem.address = fileName;
                 profileItem.remarks = "clash_custom";
             }
@@ -1132,7 +1252,7 @@ namespace v2rayN.Handler
 
             if (isSub && !Utils.IsNullOrEmpty(subid))
             {
-                RemoveServerViaSubid(ref config, subid, isSub);
+                RemoveServerViaSubid(config, subid, isSub);
             }
             if (isSub && lstOriSub?.Count == 1)
             {
@@ -1146,7 +1266,7 @@ namespace v2rayN.Handler
                 return -1;
             }
 
-            if (AddCustomServer(ref config, profileItem, true) == 0)
+            if (AddCustomServer(config, profileItem, true) == 0)
             {
                 return 1;
             }
@@ -1156,7 +1276,7 @@ namespace v2rayN.Handler
             }
         }
 
-        private static int AddBatchServers4SsSIP008(ref Config config, string clipboardData, string subid, bool isSub, List<ProfileItem> lstOriSub)
+        private static int AddBatchServers4SsSIP008(Config config, string clipboardData, string subid, bool isSub, List<ProfileItem> lstOriSub)
         {
             if (Utils.IsNullOrEmpty(clipboardData))
             {
@@ -1165,14 +1285,14 @@ namespace v2rayN.Handler
 
             if (isSub && !Utils.IsNullOrEmpty(subid))
             {
-                RemoveServerViaSubid(ref config, subid, isSub);
+                RemoveServerViaSubid(config, subid, isSub);
             }
 
             //SsSIP008
-            var lstSsServer = Utils.FromJson<List<SsServer>>(clipboardData);
+            var lstSsServer = JsonUtils.Deserialize<List<SsServer>>(clipboardData);
             if (lstSsServer?.Count <= 0)
             {
-                var ssSIP008 = Utils.FromJson<SsSIP008>(clipboardData);
+                var ssSIP008 = JsonUtils.Deserialize<SsSIP008>(clipboardData);
                 if (ssSIP008?.servers?.Count > 0)
                 {
                     lstSsServer = ssSIP008.servers;
@@ -1195,7 +1315,7 @@ namespace v2rayN.Handler
                     };
                     ssItem.subid = subid;
                     ssItem.isSub = isSub;
-                    if (AddShadowsocksServer(ref config, ssItem) == 0)
+                    if (AddShadowsocksServer(config, ssItem) == 0)
                     {
                         counter++;
                     }
@@ -1207,7 +1327,7 @@ namespace v2rayN.Handler
             return -1;
         }
 
-        public static int AddBatchServers(ref Config config, string clipboardData, string subid, bool isSub)
+        public static int AddBatchServers(Config config, string clipboardData, string subid, bool isSub)
         {
             List<ProfileItem>? lstOriSub = null;
             if (isSub && !Utils.IsNullOrEmpty(subid))
@@ -1218,26 +1338,26 @@ namespace v2rayN.Handler
             var counter = 0;
             if (Utils.IsBase64String(clipboardData))
             {
-                counter = AddBatchServers(ref config, Utils.Base64Decode(clipboardData), subid, isSub, lstOriSub);
+                counter = AddBatchServers(config, Utils.Base64Decode(clipboardData), subid, isSub, lstOriSub);
             }
             if (counter < 1)
             {
-                counter = AddBatchServers(ref config, clipboardData, subid, isSub, lstOriSub);
+                counter = AddBatchServers(config, clipboardData, subid, isSub, lstOriSub);
             }
             if (counter < 1)
             {
-                counter = AddBatchServers(ref config, Utils.Base64Decode(clipboardData), subid, isSub, lstOriSub);
+                counter = AddBatchServers(config, Utils.Base64Decode(clipboardData), subid, isSub, lstOriSub);
             }
 
             if (counter < 1)
             {
-                counter = AddBatchServers4SsSIP008(ref config, clipboardData, subid, isSub, lstOriSub);
+                counter = AddBatchServers4SsSIP008(config, clipboardData, subid, isSub, lstOriSub);
             }
 
             //maybe other sub
             if (counter < 1)
             {
-                counter = AddBatchServers4Custom(ref config, clipboardData, subid, isSub, lstOriSub);
+                counter = AddBatchServers4Custom(config, clipboardData, subid, isSub, lstOriSub);
             }
 
             return counter;
@@ -1253,7 +1373,7 @@ namespace v2rayN.Handler
         /// <param name="config"></param>
         /// <param name="url"></param>
         /// <returns></returns>
-        public static int AddSubItem(ref Config config, string url)
+        public static int AddSubItem(Config config, string url)
         {
             //already exists
             if (SqliteHelper.Instance.Table<SubItem>().Where(e => e.url == url).Count() > 0)
@@ -1268,10 +1388,10 @@ namespace v2rayN.Handler
                 url = url
             };
 
-            return AddSubItem(ref config, subItem);
+            return AddSubItem(config, subItem);
         }
 
-        public static int AddSubItem(ref Config config, SubItem subItem)
+        public static int AddSubItem(Config config, SubItem subItem)
         {
             if (Utils.IsNullOrEmpty(subItem.id))
             {
@@ -1303,7 +1423,7 @@ namespace v2rayN.Handler
         /// <param name="config"></param>
         /// <param name="subid"></param>
         /// <returns></returns>
-        public static int RemoveServerViaSubid(ref Config config, string subid, bool isSub)
+        public static int RemoveServerViaSubid(Config config, string subid, bool isSub)
         {
             if (Utils.IsNullOrEmpty(subid))
             {
@@ -1326,7 +1446,7 @@ namespace v2rayN.Handler
             return 0;
         }
 
-        public static int DeleteSubItem(ref Config config, string id)
+        public static int DeleteSubItem(Config config, string id)
         {
             var item = LazyConfig.Instance.GetSubItem(id);
             if (item is null)
@@ -1334,7 +1454,7 @@ namespace v2rayN.Handler
                 return 0;
             }
             SqliteHelper.Instance.Delete(item);
-            RemoveServerViaSubid(ref config, id, false);
+            RemoveServerViaSubid(config, id, false);
 
             return 0;
         }
@@ -1354,7 +1474,7 @@ namespace v2rayN.Handler
 
         #region Routing
 
-        public static int SaveRoutingItem(ref Config config, RoutingItem item)
+        public static int SaveRoutingItem(Config config, RoutingItem item)
         {
             if (Utils.IsNullOrEmpty(item.id))
             {
@@ -1384,7 +1504,7 @@ namespace v2rayN.Handler
                 return -1;
             }
 
-            var lstRules = Utils.FromJson<List<RulesItem>>(clipboardData);
+            var lstRules = JsonUtils.Deserialize<List<RulesItem>>(clipboardData);
             if (lstRules == null)
             {
                 return -1;
@@ -1395,7 +1515,7 @@ namespace v2rayN.Handler
                 item.id = Utils.GetGUID(false);
             }
             routingItem.ruleNum = lstRules.Count;
-            routingItem.ruleSet = Utils.ToJson(lstRules, false);
+            routingItem.ruleSet = JsonUtils.Serialize(lstRules, false);
 
             if (Utils.IsNullOrEmpty(routingItem.id))
             {
@@ -1434,7 +1554,7 @@ namespace v2rayN.Handler
                         {
                             return 0;
                         }
-                        var item = Utils.DeepCopy(rules[index]);
+                        var item = JsonUtils.DeepCopy(rules[index]);
                         rules.RemoveAt(index);
                         rules.Insert(0, item);
 
@@ -1446,7 +1566,7 @@ namespace v2rayN.Handler
                         {
                             return 0;
                         }
-                        var item = Utils.DeepCopy(rules[index]);
+                        var item = JsonUtils.DeepCopy(rules[index]);
                         rules.RemoveAt(index);
                         rules.Insert(index - 1, item);
 
@@ -1459,7 +1579,7 @@ namespace v2rayN.Handler
                         {
                             return 0;
                         }
-                        var item = Utils.DeepCopy(rules[index]);
+                        var item = JsonUtils.DeepCopy(rules[index]);
                         rules.RemoveAt(index);
                         rules.Insert(index + 1, item);
 
@@ -1471,7 +1591,7 @@ namespace v2rayN.Handler
                         {
                             return 0;
                         }
-                        var item = Utils.DeepCopy(rules[index]);
+                        var item = JsonUtils.DeepCopy(rules[index]);
                         rules.RemoveAt(index);
                         rules.Add(item);
 
@@ -1480,7 +1600,7 @@ namespace v2rayN.Handler
                 case EMove.Position:
                     {
                         var removeItem = rules[index];
-                        var item = Utils.DeepCopy(rules[index]);
+                        var item = JsonUtils.DeepCopy(rules[index]);
                         rules.Insert(pos, item);
                         rules.Remove(removeItem);
                         break;
@@ -1489,7 +1609,7 @@ namespace v2rayN.Handler
             return 0;
         }
 
-        public static int SetDefaultRouting(ref Config config, RoutingItem routingItem)
+        public static int SetDefaultRouting(Config config, RoutingItem routingItem)
         {
             if (SqliteHelper.Instance.Table<RoutingItem>().Where(t => t.id == routingItem.id).Count() > 0)
             {
@@ -1501,20 +1621,20 @@ namespace v2rayN.Handler
             return 0;
         }
 
-        public static RoutingItem GetDefaultRouting(ref Config config)
+        public static RoutingItem GetDefaultRouting(Config config)
         {
             var item = LazyConfig.Instance.GetRoutingItem(config.routingBasicItem.routingIndexId);
             if (item is null)
             {
                 var item2 = SqliteHelper.Instance.Table<RoutingItem>().FirstOrDefault(t => t.locked == false);
-                SetDefaultRouting(ref config, item2);
+                SetDefaultRouting(config, item2);
                 return item2;
             }
 
             return item;
         }
 
-        public static int InitBuiltinRouting(ref Config config, bool blImportAdvancedRules = false)
+        public static int InitBuiltinRouting(Config config, bool blImportAdvancedRules = false)
         {
             var items = LazyConfig.Instance.RoutingItems();
             if (blImportAdvancedRules || items.Count <= 0)
@@ -1549,11 +1669,11 @@ namespace v2rayN.Handler
 
                 if (!blImportAdvancedRules)
                 {
-                    SetDefaultRouting(ref config, item2);
+                    SetDefaultRouting(config, item2);
                 }
             }
 
-            if (GetLockedRoutingItem(ref config) == null)
+            if (GetLockedRoutingItem(config) == null)
             {
                 var item1 = new RoutingItem()
                 {
@@ -1566,7 +1686,7 @@ namespace v2rayN.Handler
             return 0;
         }
 
-        public static RoutingItem GetLockedRoutingItem(ref Config config)
+        public static RoutingItem GetLockedRoutingItem(Config config)
         {
             return SqliteHelper.Instance.Table<RoutingItem>().FirstOrDefault(it => it.locked == true);
         }
