@@ -35,11 +35,11 @@ namespace v2rayN.Handler
                     }
                     if (Utils.IsNullOrEmpty(fileName))
                     {
-                        content = Utils.ToJson(singboxConfig);
+                        content = JsonUtils.Serialize(singboxConfig);
                     }
                     else
                     {
-                        Utils.ToJsonFile(singboxConfig, fileName, false);
+                        JsonUtils.ToFile(singboxConfig, fileName, false);
                     }
                 }
                 else
@@ -51,17 +51,17 @@ namespace v2rayN.Handler
                     }
                     if (Utils.IsNullOrEmpty(fileName))
                     {
-                        content = Utils.ToJson(v2rayConfig);
+                        content = JsonUtils.Serialize(v2rayConfig);
                     }
                     else
                     {
-                        Utils.ToJsonFile(v2rayConfig, fileName, false);
+                        JsonUtils.ToFile(v2rayConfig, fileName, false);
                     }
                 }
             }
             catch (Exception ex)
             {
-                Utils.SaveLog("GenerateClientConfig", ex);
+                Logging.SaveLog("GenerateClientConfig", ex);
                 msg = ResUI.FailedGenDefaultConfiguration;
                 return -1;
             }
@@ -80,6 +80,7 @@ namespace v2rayN.Handler
 
                 if (File.Exists(fileName))
                 {
+                    File.SetAttributes(fileName, FileAttributes.Normal); //If the file has a read-only attribute, direct deletion will fail
                     File.Delete(fileName);
                 }
 
@@ -94,6 +95,7 @@ namespace v2rayN.Handler
                     return -1;
                 }
                 File.Copy(addressFileName, fileName);
+                File.SetAttributes(fileName, FileAttributes.Normal); //Copy will keep the attributes of addressFileName, so we need to add write permissions to fileName just in case of addressFileName is a read-only file.
 
                 //check again
                 if (!File.Exists(fileName))
@@ -117,6 +119,7 @@ namespace v2rayN.Handler
 
                         case ECoreType.clash:
                         case ECoreType.clash_meta:
+                        case ECoreType.mihomo:
                             //remove the original
                             var indexPort = fileContent.FindIndex(t => t.Contains("port:"));
                             if (indexPort >= 0)
@@ -140,17 +143,32 @@ namespace v2rayN.Handler
             }
             catch (Exception ex)
             {
-                Utils.SaveLog("GenerateClientCustomConfig", ex);
+                Logging.SaveLog("GenerateClientCustomConfig", ex);
                 msg = ResUI.FailedGenDefaultConfiguration;
                 return -1;
             }
             return 0;
         }
 
-        public static string GenerateClientSpeedtestConfigString(Config config, List<ServerTestItem> selecteds, out string msg)
+        public static int GenerateClientSpeedtestConfig(Config config, string fileName, List<ServerTestItem> selecteds, ECoreType coreType, out string msg)
         {
-            var coreConfigV2ray = new CoreConfigV2ray(config);
-            return coreConfigV2ray.GenerateClientSpeedtestConfigString(selecteds, out msg);
+            if (coreType == ECoreType.sing_box)
+            {
+                if ((new CoreConfigSingbox(config)).GenerateClientSpeedtestConfig(selecteds, out SingboxConfig? singboxConfig, out msg) != 0)
+                {
+                    return -1;
+                }
+                JsonUtils.ToFile(singboxConfig, fileName, false);
+            }
+            else
+            {
+                if ((new CoreConfigV2ray(config)).GenerateClientSpeedtestConfig(selecteds, out V2rayConfig? v2rayConfig, out msg) != 0)
+                {
+                    return -1;
+                }
+                JsonUtils.ToFile(v2rayConfig, fileName, false);
+            }
+            return 0;
         }
     }
 }
