@@ -1,4 +1,4 @@
-﻿using v2rayN.Base;
+﻿using System.Runtime.Intrinsics.X86;
 using v2rayN.Mode;
 
 namespace v2rayN.Handler
@@ -129,6 +129,15 @@ namespace v2rayN.Handler
             return SqliteHelper.Instance.Table<ProfileItem>().FirstOrDefault(it => it.indexId == indexId);
         }
 
+        public ProfileItem? GetProfileItemViaRemarks(string remarks)
+        {
+            if (Utils.IsNullOrEmpty(remarks))
+            {
+                return null;
+            }
+            return SqliteHelper.Instance.Table<ProfileItem>().FirstOrDefault(it => it.remarks == remarks);
+        }
+
         public List<RoutingItem> RoutingItems()
         {
             return SqliteHelper.Instance.Table<RoutingItem>().Where(it => it.locked == false).OrderBy(t => t.sort).ToList();
@@ -155,16 +164,19 @@ namespace v2rayN.Handler
 
         public List<string> GetShadowsocksSecuritys(ProfileItem profileItem)
         {
-            if (GetCoreType(profileItem, EConfigType.Shadowsocks) == ECoreType.v2fly)
+            var coreType = GetCoreType(profileItem, EConfigType.Shadowsocks);
+            switch (coreType)
             {
-                return Global.ssSecuritys;
-            }
-            if (GetCoreType(profileItem, EConfigType.Shadowsocks) == ECoreType.Xray)
-            {
-                return Global.ssSecuritysInXray;
-            }
+                case ECoreType.v2fly:
+                    return Global.SsSecuritys;
 
-            return Global.ssSecuritysInSagerNet;
+                case ECoreType.Xray:
+                    return Global.SsSecuritysInXray;
+
+                case ECoreType.sing_box:
+                    return Global.SsSecuritysInSingbox;
+            }
+            return Global.SsSecuritysInSagerNet;
         }
 
         public ECoreType GetCoreType(ProfileItem profileItem, EConfigType eConfigType)
@@ -195,13 +207,13 @@ namespace v2rayN.Handler
             return coreInfos!.FirstOrDefault(t => t.coreType == coreType);
         }
 
-        public List<CoreInfo>? GetCoreInfos()
+        public List<CoreInfo> GetCoreInfos()
         {
             if (coreInfos == null)
             {
                 InitCoreInfo();
             }
-            return coreInfos;
+            return coreInfos!;
         }
 
         private void InitCoreInfo()
@@ -212,7 +224,7 @@ namespace v2rayN.Handler
             {
                 coreType = ECoreType.v2rayN,
                 coreUrl = Global.NUrl,
-                coreReleaseApiUrl = Global.NUrl.Replace(Global.githubUrl, Global.githubApiUrl),
+                coreReleaseApiUrl = Global.NUrl.Replace(Global.GithubUrl, Global.GithubApiUrl),
                 coreDownloadUrl32 = Global.NUrl + "/download/{0}/v2rayN-32.zip",
                 coreDownloadUrl64 = Global.NUrl + "/download/{0}/v2rayN.zip",
                 coreDownloadUrlArm64 = Global.NUrl + "/download/{0}/v2rayN-arm64.zip"
@@ -223,11 +235,11 @@ namespace v2rayN.Handler
                 coreType = ECoreType.v2fly,
                 coreExes = new List<string> { "wv2ray", "v2ray" },
                 arguments = "",
-                coreUrl = Global.v2flyCoreUrl,
-                coreReleaseApiUrl = Global.v2flyCoreUrl.Replace(Global.githubUrl, Global.githubApiUrl),
-                coreDownloadUrl32 = Global.v2flyCoreUrl + "/download/{0}/v2ray-windows-{1}.zip",
-                coreDownloadUrl64 = Global.v2flyCoreUrl + "/download/{0}/v2ray-windows-{1}.zip",
-                coreDownloadUrlArm64 = Global.v2flyCoreUrl + "/download/{0}/v2ray-windows-{1}.zip",
+                coreUrl = Global.V2flyCoreUrl,
+                coreReleaseApiUrl = Global.V2flyCoreUrl.Replace(Global.GithubUrl, Global.GithubApiUrl),
+                coreDownloadUrl32 = Global.V2flyCoreUrl + "/download/{0}/v2ray-windows-{1}.zip",
+                coreDownloadUrl64 = Global.V2flyCoreUrl + "/download/{0}/v2ray-windows-{1}.zip",
+                coreDownloadUrlArm64 = Global.V2flyCoreUrl + "/download/{0}/v2ray-windows-{1}.zip",
                 match = "V2Ray",
                 versionArg = "-version",
                 redirectInfo = true,
@@ -239,7 +251,7 @@ namespace v2rayN.Handler
                 coreExes = new List<string> { "SagerNet", "v2ray" },
                 arguments = "run",
                 coreUrl = Global.SagerNetCoreUrl,
-                coreReleaseApiUrl = Global.SagerNetCoreUrl.Replace(Global.githubUrl, Global.githubApiUrl),
+                coreReleaseApiUrl = Global.SagerNetCoreUrl.Replace(Global.GithubUrl, Global.GithubApiUrl),
                 coreDownloadUrl32 = Global.SagerNetCoreUrl + "/download/{0}/v2ray-windows-{1}.zip",
                 coreDownloadUrl64 = Global.SagerNetCoreUrl + "/download/{0}/v2ray-windows-{1}.zip",
                 coreDownloadUrlArm64 = Global.SagerNetCoreUrl + "/download/{0}/v2ray-windows-{1}.zip",
@@ -252,12 +264,12 @@ namespace v2rayN.Handler
             {
                 coreType = ECoreType.v2fly_v5,
                 coreExes = new List<string> { "v2ray" },
-                arguments = "run",
-                coreUrl = Global.v2flyCoreUrl,
-                coreReleaseApiUrl = Global.v2flyCoreUrl.Replace(Global.githubUrl, Global.githubApiUrl),
-                coreDownloadUrl32 = Global.v2flyCoreUrl + "/download/{0}/v2ray-windows-{1}.zip",
-                coreDownloadUrl64 = Global.v2flyCoreUrl + "/download/{0}/v2ray-windows-{1}.zip",
-                coreDownloadUrlArm64 = Global.v2flyCoreUrl + "/download/{0}/v2ray-windows-{1}.zip",
+                arguments = "run -c config.json -format jsonv5",
+                coreUrl = Global.V2flyCoreUrl,
+                coreReleaseApiUrl = Global.V2flyCoreUrl.Replace(Global.GithubUrl, Global.GithubApiUrl),
+                coreDownloadUrl32 = Global.V2flyCoreUrl + "/download/{0}/v2ray-windows-{1}.zip",
+                coreDownloadUrl64 = Global.V2flyCoreUrl + "/download/{0}/v2ray-windows-{1}.zip",
+                coreDownloadUrlArm64 = Global.V2flyCoreUrl + "/download/{0}/v2ray-windows-{1}.zip",
                 match = "V2Ray",
                 versionArg = "version",
                 redirectInfo = true,
@@ -267,12 +279,12 @@ namespace v2rayN.Handler
             {
                 coreType = ECoreType.Xray,
                 coreExes = new List<string> { "xray", "wxray" },
-                arguments = "",
-                coreUrl = Global.xrayCoreUrl,
-                coreReleaseApiUrl = Global.xrayCoreUrl.Replace(Global.githubUrl, Global.githubApiUrl),
-                coreDownloadUrl32 = Global.xrayCoreUrl + "/download/{0}/Xray-windows-{1}.zip",
-                coreDownloadUrl64 = Global.xrayCoreUrl + "/download/{0}/Xray-windows-{1}.zip",
-                coreDownloadUrlArm64 = Global.xrayCoreUrl + "/download/{0}/Xray-windows-{1}.zip",
+                arguments = "run {0}",
+                coreUrl = Global.XrayCoreUrl,
+                coreReleaseApiUrl = Global.XrayCoreUrl.Replace(Global.GithubUrl, Global.GithubApiUrl),
+                coreDownloadUrl32 = Global.XrayCoreUrl + "/download/{0}/Xray-windows-{1}.zip",
+                coreDownloadUrl64 = Global.XrayCoreUrl + "/download/{0}/Xray-windows-{1}.zip",
+                coreDownloadUrlArm64 = Global.XrayCoreUrl + "/download/{0}/Xray-windows-{1}.zip",
                 match = "Xray",
                 versionArg = "-version",
                 redirectInfo = true,
@@ -283,11 +295,11 @@ namespace v2rayN.Handler
                 coreType = ECoreType.clash,
                 coreExes = new List<string> { "clash-windows-amd64-v3", "clash-windows-amd64", "clash-windows-386", "clash" },
                 arguments = "-f config.json",
-                coreUrl = Global.clashCoreUrl,
-                coreReleaseApiUrl = Global.clashCoreUrl.Replace(Global.githubUrl, Global.githubApiUrl),
-                coreDownloadUrl32 = Global.clashCoreUrl + "/download/{0}/clash-windows-386-{0}.zip",
-                coreDownloadUrl64 = Global.clashCoreUrl + "/download/{0}/clash-windows-amd64-{0}.zip",
-                coreDownloadUrlArm64 = Global.clashCoreUrl + "/download/{0}/clash-windows-arm64-{0}.zip",
+                coreUrl = Global.ClashCoreUrl,
+                coreReleaseApiUrl = Global.ClashCoreUrl.Replace(Global.GithubUrl, Global.GithubApiUrl),
+                coreDownloadUrl32 = Global.ClashCoreUrl + "/download/{0}/clash-windows-386-{0}.zip",
+                coreDownloadUrl64 = Global.ClashCoreUrl + "/download/{0}/clash-windows-amd64-{0}.zip",
+                coreDownloadUrlArm64 = Global.ClashCoreUrl + "/download/{0}/clash-windows-arm64-{0}.zip",
                 match = "v",
                 versionArg = "-v",
                 redirectInfo = true,
@@ -298,13 +310,24 @@ namespace v2rayN.Handler
                 coreType = ECoreType.clash_meta,
                 coreExes = new List<string> { "Clash.Meta-windows-amd64-compatible", "Clash.Meta-windows-amd64", "Clash.Meta-windows-386", "Clash.Meta", "clash" },
                 arguments = "-f config.json",
-                coreUrl = Global.clashMetaCoreUrl,
-                coreReleaseApiUrl = Global.clashMetaCoreUrl.Replace(Global.githubUrl, Global.githubApiUrl),
-                coreDownloadUrl32 = Global.clashMetaCoreUrl + "/download/{0}/Clash.Meta-windows-386-{0}.zip",
-                coreDownloadUrl64 = Global.clashMetaCoreUrl + "/download/{0}/Clash.Meta-windows-amd64-compatible-{0}.zip",
-                coreDownloadUrlArm64 = Global.clashMetaCoreUrl + "/download/{0}/Clash.Meta-windows-arm64-{0}.zip",
+                coreUrl = Global.ClashMetaCoreUrl,
+                coreReleaseApiUrl = Global.ClashMetaCoreUrl.Replace(Global.GithubUrl, Global.GithubApiUrl),
+                coreDownloadUrl32 = Global.ClashMetaCoreUrl + "/download/{0}/Clash.Meta-windows-386-{0}.zip",
+                coreDownloadUrl64 = Global.ClashMetaCoreUrl + "/download/{0}/Clash.Meta-windows-amd64-compatible-{0}.zip",
+                coreDownloadUrlArm64 = Global.ClashMetaCoreUrl + "/download/{0}/Clash.Meta-windows-arm64-{0}.zip",
                 match = "v",
                 versionArg = "-v",
+                redirectInfo = true,
+            });
+
+            coreInfos.Add(new CoreInfo
+            {
+                coreType = ECoreType.mihomo,
+                coreExes = new List<string> { $"mihomo-windows-amd64{(Avx2.X64.IsSupported ? "" : "-compatible")}", "mihomo-windows-amd64-compatible", "mihomo-windows-amd64", "mihomo-windows-386", "mihomo", "clash" },
+                arguments = "-f config.yaml",
+                coreUrl = Global.MihomoCoreUrl,
+                coreReleaseApiUrl = Global.MihomoCoreUrl.Replace(Global.GithubUrl, Global.GithubApiUrl),
+                match = "Mihomo",
                 redirectInfo = true,
             });
 
@@ -313,11 +336,11 @@ namespace v2rayN.Handler
                 coreType = ECoreType.hysteria,
                 coreExes = new List<string> { "hysteria-windows-amd64", "hysteria-windows-386", "hysteria" },
                 arguments = "",
-                coreUrl = Global.hysteriaCoreUrl,
-                coreReleaseApiUrl = Global.hysteriaCoreUrl.Replace(Global.githubUrl, Global.githubApiUrl),
-                coreDownloadUrl32 = Global.hysteriaCoreUrl + "/download/{0}/hysteria-windows-386.exe",
-                coreDownloadUrl64 = Global.hysteriaCoreUrl + "/download/{0}/hysteria-windows-amd64.exe",
-                coreDownloadUrlArm64 = Global.hysteriaCoreUrl + "/download/{0}/hysteria-windows-arm64.exe",
+                coreUrl = Global.HysteriaCoreUrl,
+                coreReleaseApiUrl = Global.HysteriaCoreUrl.Replace(Global.GithubUrl, Global.GithubApiUrl),
+                coreDownloadUrl32 = Global.HysteriaCoreUrl + "/download/{0}/hysteria-windows-386.exe",
+                coreDownloadUrl64 = Global.HysteriaCoreUrl + "/download/{0}/hysteria-windows-amd64.exe",
+                coreDownloadUrlArm64 = Global.HysteriaCoreUrl + "/download/{0}/hysteria-windows-arm64.exe",
                 redirectInfo = true,
             });
 
@@ -326,7 +349,7 @@ namespace v2rayN.Handler
                 coreType = ECoreType.naiveproxy,
                 coreExes = new List<string> { "naiveproxy", "naive" },
                 arguments = "config.json",
-                coreUrl = Global.naiveproxyCoreUrl,
+                coreUrl = Global.NaiveproxyCoreUrl,
                 redirectInfo = false,
             });
 
@@ -335,7 +358,7 @@ namespace v2rayN.Handler
                 coreType = ECoreType.tuic,
                 coreExes = new List<string> { "tuic-client", "tuic" },
                 arguments = "-c config.json",
-                coreUrl = Global.tuicCoreUrl,
+                coreUrl = Global.TuicCoreUrl,
                 redirectInfo = true,
             });
 
@@ -343,13 +366,13 @@ namespace v2rayN.Handler
             {
                 coreType = ECoreType.sing_box,
                 coreExes = new List<string> { "sing-box-client", "sing-box" },
-                arguments = "run{0}",
-                coreUrl = Global.singboxCoreUrl,
+                arguments = "run {0} --disable-color",
+                coreUrl = Global.SingboxCoreUrl,
                 redirectInfo = true,
-                coreReleaseApiUrl = Global.singboxCoreUrl.Replace(Global.githubUrl, Global.githubApiUrl),
-                coreDownloadUrl32 = Global.singboxCoreUrl + "/download/{0}/sing-box-{1}-windows-386.zip",
-                coreDownloadUrl64 = Global.singboxCoreUrl + "/download/{0}/sing-box-{1}-windows-amd64.zip",
-                coreDownloadUrlArm64 = Global.singboxCoreUrl + "/download/{0}/sing-box-{1}-windows-arm64.zip",
+                coreReleaseApiUrl = Global.SingboxCoreUrl.Replace(Global.GithubUrl, Global.GithubApiUrl),
+                coreDownloadUrl32 = Global.SingboxCoreUrl + "/download/{0}/sing-box-{1}-windows-386.zip",
+                coreDownloadUrl64 = Global.SingboxCoreUrl + "/download/{0}/sing-box-{1}-windows-amd64.zip",
+                coreDownloadUrlArm64 = Global.SingboxCoreUrl + "/download/{0}/sing-box-{1}-windows-arm64.zip",
                 match = "sing-box",
                 versionArg = "version",
             });
@@ -359,7 +382,20 @@ namespace v2rayN.Handler
                 coreType = ECoreType.juicity,
                 coreExes = new List<string> { "juicity-client", "juicity" },
                 arguments = "run -c config.json",
-                coreUrl = Global.juicityCoreUrl
+                coreUrl = Global.JuicityCoreUrl
+            });
+
+            coreInfos.Add(new CoreInfo
+            {
+                coreType = ECoreType.hysteria2,
+                coreExes = new List<string> { "hysteria-windows-amd64", "hysteria-windows-386", "hysteria" },
+                arguments = "",
+                coreUrl = Global.HysteriaCoreUrl,
+                coreReleaseApiUrl = Global.HysteriaCoreUrl.Replace(Global.GithubUrl, Global.GithubApiUrl),
+                coreDownloadUrl32 = Global.HysteriaCoreUrl + "/download/{0}/hysteria-windows-386.exe",
+                coreDownloadUrl64 = Global.HysteriaCoreUrl + "/download/{0}/hysteria-windows-amd64.exe",
+                coreDownloadUrlArm64 = Global.HysteriaCoreUrl + "/download/{0}/hysteria-windows-arm64.exe",
+                redirectInfo = true,
             });
         }
 
