@@ -1,6 +1,5 @@
 ﻿using System.Collections.Concurrent;
 using System.Reactive.Linq;
-using v2rayN.Base;
 using v2rayN.Mode;
 
 namespace v2rayN.Handler
@@ -51,7 +50,7 @@ namespace v2rayN.Handler
             }
         }
 
-        private void AddProfileEx(string indexId, ref ProfileExItem profileEx)
+        private void AddProfileEx(string indexId, ref ProfileExItem? profileEx)
         {
             profileEx = new()
             {
@@ -82,7 +81,7 @@ namespace v2rayN.Handler
             }
             catch (Exception ex)
             {
-                Utils.SaveLog(ex.Message, ex);
+                Logging.SaveLog(ex.Message, ex);
             }
         }
 
