@@ -1,10 +1,5 @@
 ﻿using PacLib;
-using System.Diagnostics;
-using System.IO;
-using System.Text;
 using v2rayN.Mode;
-using v2rayN.Properties;
-using v2rayN.Tool;
 
 namespace v2rayN.Handler
 {
@@ -20,7 +15,6 @@ namespace v2rayN.Handler
         //  <proxy-server><CR-LF>
         //  <bypass-list><CR-LF>
         //  <pac-url>
-        private static SysproxyConfig? _userSettings = null;
 
         private enum RET_ERRORS : int
         {
@@ -34,15 +28,6 @@ namespace v2rayN.Handler
 
         static SysProxyHandle()
         {
-            try
-            {
-                FileManager.UncompressFile(Utils.GetTempPath("sysproxy.exe"),
-                    Environment.Is64BitOperatingSystem ? Resources.sysproxy64_exe : Resources.sysproxy_exe);
-            }
-            catch (IOException ex)
-            {
-                Utils.SaveLog(ex.Message, ex);
-            }
         }
 
         public static bool UpdateSysProxy(Config config, bool forceDisable)
@@ -79,13 +64,11 @@ namespace v2rayN.Handler
                             .Replace("{http_port}", port.ToString())
                             .Replace("{socks_port}", portSocks.ToString());
                     }
-                    ProxySetting.SetProxy(strProxy, strExceptions, 2);
-                    SetIEProxy(true, strProxy, strExceptions);
+                    ProxySetting.SetProxy(strProxy, strExceptions, 2); // set a named proxy
                 }
                 else if (type == ESysProxyType.ForcedClear)
                 {
-                    ProxySetting.UnsetProxy();
-                    ResetIEProxy();
+                    ProxySetting.UnsetProxy(); // set to no proxy
                 }
                 else if (type == ESysProxyType.Unchanged)
                 {
@@ -93,9 +76,8 @@ namespace v2rayN.Handler
                 else if (type == ESysProxyType.Pac)
                 {
                     PacHandler.Start(Utils.GetConfigPath(), port, portPac);
-                    var strProxy = $"{Global.httpProtocol}{Global.Loopback}:{portPac}/pac?t={DateTime.Now.Ticks}";
-                    ProxySetting.SetProxy(strProxy, "", 4);
-                    SetIEProxy(false, strProxy, "");
+                    var strProxy = $"{Global.HttpProtocol}{Global.Loopback}:{portPac}/pac?t={DateTime.Now.Ticks}";
+                    ProxySetting.SetProxy(strProxy, "", 4); // use pac script url for auto-config proxy
                 }
 
                 if (type != ESysProxyType.Pac)
@@ -105,7 +87,7 @@ namespace v2rayN.Handler
             }
             catch (Exception ex)
             {
-                Utils.SaveLog(ex.Message, ex);
+                Logging.SaveLog(ex.Message, ex);
             }
             return true;
         }
@@ -120,103 +102,6 @@ namespace v2rayN.Handler
             catch
             {
             }
-        }
-
-        public static void SetIEProxy(bool global, string strProxy, string strExceptions)
-        {
-            string arguments = global
-                ? $"global {strProxy} {strExceptions}"
-                : $"pac {strProxy}";
-
-            ExecSysproxy(arguments);
-        }
-
-        // set system proxy to 1 (null) (null) (null)
-        public static void ResetIEProxy()
-        {
-            ExecSysproxy("set 1 - - -");
-        }
-
-        private static void ExecSysproxy(string arguments)
-        {
-            // using event to avoid hanging when redirect standard output/error
-            // ref: https://stackoverflow.com/questions/139593/processstartinfo-hanging-on-waitforexit-why
-            // and http://blog.csdn.net/zhangweixing0/article/details/7356841
-            using AutoResetEvent outputWaitHandle = new(false);
-            using AutoResetEvent errorWaitHandle = new(false);
-            using Process process = new();
-
-            // Configure the process using the StartInfo properties.
-            process.StartInfo.FileName = Utils.GetTempPath("sysproxy.exe");
-            process.StartInfo.Arguments = arguments;
-            process.StartInfo.WorkingDirectory = Utils.GetTempPath();
-            process.StartInfo.WindowStyle = ProcessWindowStyle.Hidden;
-            process.StartInfo.UseShellExecute = false;
-            process.StartInfo.RedirectStandardError = true;
-            process.StartInfo.RedirectStandardOutput = true;
-
-            // Need to provide encoding info, or output/error strings we got will be wrong.
-            process.StartInfo.StandardOutputEncoding = Encoding.Unicode;
-            process.StartInfo.StandardErrorEncoding = Encoding.Unicode;
-
-            process.StartInfo.CreateNoWindow = true;
-
-            StringBuilder output = new(1024);
-            StringBuilder error = new(1024);
-
-            process.OutputDataReceived += (sender, e) =>
-            {
-                if (e.Data == null)
-                {
-                    outputWaitHandle.Set();
-                }
-                else
-                {
-                    output.AppendLine(e.Data);
-                }
-            };
-            process.ErrorDataReceived += (sender, e) =>
-            {
-                if (e.Data == null)
-                {
-                    errorWaitHandle.Set();
-                }
-                else
-                {
-                    error.AppendLine(e.Data);
-                }
-            };
-            try
-            {
-                process.Start();
-
-                process.BeginErrorReadLine();
-                process.BeginOutputReadLine();
-
-                process.WaitForExit();
-            }
-            catch (System.ComponentModel.Win32Exception e)
-            {
-                // log the arguments
-                throw new Exception(process.StartInfo.Arguments);
-            }
-            string stderr = error.ToString();
-            string stdout = output.ToString();
-
-            int exitCode = process.ExitCode;
-            if (exitCode != (int)RET_ERRORS.RET_NO_ERROR)
-            {
-                throw new Exception(stderr);
-            }
-
-            //if (arguments == "query")
-            //{
-            //    if (stdout.IsNullOrWhiteSpace() || stdout.IsNullOrEmpty())
-            //    {
-            //        throw new Exception("failed to query wininet settings");
-            //    }
-            //    _queryStr = stdout;
-            //}
         }
     }
 }
