@@ -6,10 +6,8 @@ using System.Runtime.InteropServices;
 using System.Text;
 using System.Text.RegularExpressions;
 using System.Windows;
-using v2rayN.Base;
 using v2rayN.Mode;
 using v2rayN.Resx;
-using v2rayN.Tool;
 
 namespace v2rayN.Handler
 {
@@ -24,11 +22,13 @@ namespace v2rayN.Handler
         {
             public bool Success;
             public string Msg;
+            public string Url;
 
-            public ResultEventArgs(bool success, string msg)
+            public ResultEventArgs(bool success, string msg, string url = "")
             {
                 Success = success;
                 Msg = msg;
+                Url = url;
             }
         }
 
@@ -83,8 +83,9 @@ namespace v2rayN.Handler
                 if (args.Success)
                 {
                     _updateFunc(false, string.Format(ResUI.MsgParsingSuccessfully, "v2rayN"));
+                    _updateFunc(false, args.Msg);
 
-                    url = args.Msg;
+                    url = args.Url;
                     _ = askToDownload(downloadHandle, url, true);
                 }
                 else
@@ -135,7 +136,9 @@ namespace v2rayN.Handler
                 if (args.Success)
                 {
                     _updateFunc(false, string.Format(ResUI.MsgParsingSuccessfully, "Core"));
-                    url = args.Msg;
+                    _updateFunc(false, args.Msg);
+
+                    url = args.Url;
                     _ = askToDownload(downloadHandle, url, true);
                 }
                 else
@@ -175,7 +178,7 @@ namespace v2rayN.Handler
                         //_updateFunc(false, $"{hashCode}{ResUI.MsgNoValidSubscription}");
                         continue;
                     }
-                    if (!url.StartsWith(Global.httpsProtocol) && !url.StartsWith(Global.httpProtocol))
+                    if (!url.StartsWith(Global.HttpsProtocol) && !url.StartsWith(Global.HttpProtocol))
                     {
                         continue;
                     }
@@ -218,7 +221,7 @@ namespace v2rayN.Handler
                     //more url
                     if (Utils.IsNullOrEmpty(item.convertTarget) && !Utils.IsNullOrEmpty(item.moreUrl.TrimEx()))
                     {
-                        if (!Utils.IsNullOrEmpty(result) && Utils.IsBase64String(result))
+                        if (!Utils.IsNullOrEmpty(result) && Utils.IsBase64String(result!))
                         {
                             result = Utils.Base64Decode(result);
                         }
@@ -242,7 +245,7 @@ namespace v2rayN.Handler
                             }
                             if (!Utils.IsNullOrEmpty(result2))
                             {
-                                if (Utils.IsBase64String(result2))
+                                if (Utils.IsBase64String(result2!))
                                 {
                                     result += Utils.Base64Decode(result2);
                                 }
@@ -266,11 +269,11 @@ namespace v2rayN.Handler
                             _updateFunc(false, $"{hashCode}{result}");
                         }
 
-                        int ret = ConfigHandler.AddBatchServers(ref config, result, id, true);
+                        int ret = ConfigHandler.AddBatchServers(config, result, id, true);
                         if (ret <= 0)
                         {
-                            Utils.SaveLog("FailedImportSubscription");
-                            Utils.SaveLog(result);
+                            Logging.SaveLog("FailedImportSubscription");
+                            Logging.SaveLog(result);
                         }
                         _updateFunc(false,
                             ret > 0
@@ -291,8 +294,8 @@ namespace v2rayN.Handler
                 await UpdateGeoFile("geosite", _config, update);
                 await UpdateGeoFile("geoip", _config, update);
 
-                await UpdateGeoFile4Singbox("geosite", _config, false, update);
-                await UpdateGeoFile4Singbox("geoip", _config, true, update);
+                //await UpdateGeoFile4Singbox("geosite", _config, false, update);
+                //await UpdateGeoFile4Singbox("geoip", _config, true, update);
             });
         }
 
@@ -322,13 +325,13 @@ namespace v2rayN.Handler
                 }
                 else
                 {
-                    Utils.SaveLog("StatusCode error: " + url);
+                    Logging.SaveLog("StatusCode error: " + url);
                     return;
                 }
             }
             catch (Exception ex)
             {
-                Utils.SaveLog(ex.Message, ex);
+                Logging.SaveLog(ex.Message, ex);
                 _updateFunc(false, ex.Message);
             }
         }
@@ -345,7 +348,7 @@ namespace v2rayN.Handler
                 foreach (string name in coreInfo.coreExes)
                 {
                     string vName = $"{name}.exe";
-                    vName = Utils.GetBinPath(vName, coreInfo.coreType);
+                    vName = Utils.GetBinPath(vName, coreInfo.coreType.ToString());
                     if (File.Exists(vName))
                     {
                         filePath = vName;
@@ -394,7 +397,7 @@ namespace v2rayN.Handler
             }
             catch (Exception ex)
             {
-                Utils.SaveLog(ex.Message, ex);
+                Logging.SaveLog(ex.Message, ex);
                 _updateFunc(false, ex.Message);
                 return new SemanticVersion("");
             }
@@ -404,16 +407,11 @@ namespace v2rayN.Handler
         {
             try
             {
-                var gitHubReleases = Utils.FromJson<List<GitHubRelease>>(gitHubReleaseApi);
-                SemanticVersion version;
-                if (preRelease)
-                {
-                    version = new SemanticVersion(gitHubReleases!.First().TagName);
-                }
-                else
-                {
-                    version = new SemanticVersion(gitHubReleases!.First(r => r.Prerelease == false).TagName);
-                }
+                var gitHubReleases = JsonUtils.Deserialize<List<GitHubRelease>>(gitHubReleaseApi);
+                var gitHubRelease = preRelease ? gitHubReleases!.First() : gitHubReleases!.First(r => r.Prerelease == false);
+                var version = new SemanticVersion(gitHubRelease!.TagName);
+                var body = gitHubRelease!.Body;
+
                 var coreInfo = LazyConfig.Instance.GetCoreInfo(type);
 
                 SemanticVersion curVersion;
@@ -427,7 +425,7 @@ namespace v2rayN.Handler
                     case ECoreType.v2fly_v5:
                         {
                             curVersion = getCoreVersion(type);
-                            message = string.Format(ResUI.IsLatestCore, curVersion.ToVersionString("v"));
+                            message = string.Format(ResUI.IsLatestCore, type, curVersion.ToVersionString("v"));
                             string osBit = "64";
                             switch (RuntimeInformation.ProcessArchitecture)
                             {
@@ -451,7 +449,7 @@ namespace v2rayN.Handler
                     case ECoreType.clash_meta:
                         {
                             curVersion = getCoreVersion(type);
-                            message = string.Format(ResUI.IsLatestCore, curVersion);
+                            message = string.Format(ResUI.IsLatestCore, type, curVersion);
                             switch (RuntimeInformation.ProcessArchitecture)
                             {
                                 case Architecture.Arm64:
@@ -472,7 +470,7 @@ namespace v2rayN.Handler
                     case ECoreType.sing_box:
                         {
                             curVersion = getCoreVersion(type);
-                            message = string.Format(ResUI.IsLatestCore, curVersion.ToVersionString("v"));
+                            message = string.Format(ResUI.IsLatestCore, type, curVersion.ToVersionString("v"));
                             switch (RuntimeInformation.ProcessArchitecture)
                             {
                                 case Architecture.Arm64:
@@ -493,7 +491,7 @@ namespace v2rayN.Handler
                     case ECoreType.v2rayN:
                         {
                             curVersion = new SemanticVersion(FileVersionInfo.GetVersionInfo(Utils.GetExePath()).FileVersion.ToString());
-                            message = string.Format(ResUI.IsLatestN, curVersion);
+                            message = string.Format(ResUI.IsLatestN, type, curVersion);
                             switch (RuntimeInformation.ProcessArchitecture)
                             {
                                 case Architecture.Arm64:
@@ -520,11 +518,11 @@ namespace v2rayN.Handler
                     return;
                 }
 
-                AbsoluteCompleted?.Invoke(this, new ResultEventArgs(true, url));
+                AbsoluteCompleted?.Invoke(this, new ResultEventArgs(true, body, url));
             }
             catch (Exception ex)
             {
-                Utils.SaveLog(ex.Message, ex);
+                Logging.SaveLog(ex.Message, ex);
                 _updateFunc(false, ex.Message);
             }
         }
@@ -553,7 +551,7 @@ namespace v2rayN.Handler
         {
             _config = config;
             _updateFunc = update;
-            var url = string.Format(Global.geoUrl, geoName);
+            var url = string.Format(Global.GeoUrl, geoName);
 
             DownloadHandle downloadHandle = new();
             downloadHandle.UpdateCompleted += (sender2, args) =>
@@ -600,10 +598,10 @@ namespace v2rayN.Handler
         {
             _config = config;
             _updateFunc = update;
-            var url = string.Format(Global.singboxGeoUrl, geoName);
+            var url = string.Format(Global.SingboxGeoUrl, geoName);
 
             DownloadHandle downloadHandle = new();
-            downloadHandle.UpdateCompleted += (sender2, args) =>
+            downloadHandle.UpdateCompleted += async (sender2, args) =>
             {
                 if (args.Success)
                 {
@@ -612,8 +610,11 @@ namespace v2rayN.Handler
 
                     try
                     {
-                        if (needStop) coreHandler?.CoreStop();
-                        Task.Delay(1000);
+                        if (needStop)
+                        {
+                            coreHandler?.CoreStop();
+                            await Task.Delay(3000);
+                        }
                         string fileName = Utils.GetTempPath(Utils.GetDownloadFileName(url));
                         if (File.Exists(fileName))
                         {
