@@ -27,7 +27,8 @@
         public UIItem uiItem { get; set; }
         public ConstItem constItem { get; set; }
         public SpeedTestItem speedTestItem { get; set; }
-        public Mux4Sbox mux4Sbox { get; set; }
+        public Mux4SboxItem mux4SboxItem { get; set; }
+        public HysteriaItem hysteriaItem { get; set; }
         public List<InItem> inbound { get; set; }
         public List<KeyEventItem> globalHotkeys { get; set; }
         public List<CoreTypeItem> coreTypeItem { get; set; }
