@@ -164,6 +164,7 @@ namespace v2rayN.Mode
         public string stack { get; set; }
         public int mtu { get; set; }
         public bool enableExInbound { get; set; }
+        public bool enableIPv6Address { get; set; } = true;
     }
 
     [Serializable]
@@ -198,12 +199,19 @@ namespace v2rayN.Mode
     }
 
     [Serializable]
-    public class Mux4Sbox
+    public class Mux4SboxItem
     {
         public string protocol { get; set; }
         public int max_connections { get; set; }
         public int min_streams { get; set; }
         public int max_streams { get; set; }
         public bool padding { get; set; }
+    }
+
+    [Serializable]
+    public class HysteriaItem
+    {
+        public int up_mbps { get; set; }
+        public int down_mbps { get; set; }
     }
 }
