@@ -3,7 +3,6 @@ using ReactiveUI.Fody.Helpers;
 using Splat;
 using System.Reactive;
 using System.Windows;
-using v2rayN.Base;
 using v2rayN.Handler;
 using v2rayN.Mode;
 using v2rayN.Resx;
@@ -37,7 +36,7 @@ namespace v2rayN.ViewModels
             }
             else
             {
-                SelectedSource = Utils.DeepCopy(profileItem);
+                SelectedSource = JsonUtils.DeepCopy(profileItem);
             }
 
             SaveCmd = ReactiveCommand.Create(() =>
@@ -127,23 +126,35 @@ namespace v2rayN.ViewModels
             switch (item.configType)
             {
                 case EConfigType.VMess:
-                    ret = ConfigHandler.AddServer(ref _config, item);
+                    ret = ConfigHandler.AddServer(_config, item);
                     break;
 
                 case EConfigType.Shadowsocks:
-                    ret = ConfigHandler.AddShadowsocksServer(ref _config, item);
+                    ret = ConfigHandler.AddShadowsocksServer(_config, item);
                     break;
 
                 case EConfigType.Socks:
-                    ret = ConfigHandler.AddSocksServer(ref _config, item);
+                    ret = ConfigHandler.AddSocksServer(_config, item);
                     break;
 
                 case EConfigType.VLESS:
-                    ret = ConfigHandler.AddVlessServer(ref _config, item);
+                    ret = ConfigHandler.AddVlessServer(_config, item);
                     break;
 
                 case EConfigType.Trojan:
-                    ret = ConfigHandler.AddTrojanServer(ref _config, item);
+                    ret = ConfigHandler.AddTrojanServer(_config, item);
+                    break;
+
+                case EConfigType.Hysteria2:
+                    ret = ConfigHandler.AddHysteria2Server(_config, item);
+                    break;
+
+                case EConfigType.Tuic:
+                    ret = ConfigHandler.AddTuicServer(_config, item);
+                    break;
+
+                case EConfigType.Wireguard:
+                    ret = ConfigHandler.AddWireguardServer(_config, item);
                     break;
             }
 
