@@ -3,7 +3,6 @@ using DynamicData.Binding;
 using MaterialDesignColors;
 using MaterialDesignColors.ColorManipulation;
 using MaterialDesignThemes.Wpf;
-using Microsoft.Win32;
 using ReactiveUI;
 using ReactiveUI.Fody.Helpers;
 using Splat;
@@ -15,11 +14,9 @@ using System.Reactive.Linq;
 using System.Text;
 using System.Windows;
 using System.Windows.Media;
-using v2rayN.Base;
 using v2rayN.Handler;
 using v2rayN.Mode;
 using v2rayN.Resx;
-using v2rayN.Tool;
 using v2rayN.Views;
 
 namespace v2rayN.ViewModels
@@ -89,6 +86,9 @@ namespace v2rayN.ViewModels
         public ReactiveCommand<Unit, Unit> AddShadowsocksServerCmd { get; }
         public ReactiveCommand<Unit, Unit> AddSocksServerCmd { get; }
         public ReactiveCommand<Unit, Unit> AddTrojanServerCmd { get; }
+        public ReactiveCommand<Unit, Unit> AddHysteria2ServerCmd { get; }
+        public ReactiveCommand<Unit, Unit> AddTuicServerCmd { get; }
+        public ReactiveCommand<Unit, Unit> AddWireguardServerCmd { get; }
         public ReactiveCommand<Unit, Unit> AddCustomServerCmd { get; }
         public ReactiveCommand<Unit, Unit> AddServerViaClipboardCmd { get; }
         public ReactiveCommand<Unit, Unit> AddServerViaScanCmd { get; }
@@ -146,12 +146,15 @@ namespace v2rayN.ViewModels
         //CheckUpdate
         public ReactiveCommand<Unit, Unit> CheckUpdateNCmd { get; }
 
-        public ReactiveCommand<Unit, Unit> CheckUpdateV2flyCoreCmd { get; }
-        public ReactiveCommand<Unit, Unit> CheckUpdateSagerNetCoreCmd { get; }
+        //public ReactiveCommand<Unit, Unit> CheckUpdateV2flyCoreCmd { get; }
+
+        //public ReactiveCommand<Unit, Unit> CheckUpdateSagerNetCoreCmd { get; }
         public ReactiveCommand<Unit, Unit> CheckUpdateXrayCoreCmd { get; }
-        public ReactiveCommand<Unit, Unit> CheckUpdateClashCoreCmd { get; }
-        public ReactiveCommand<Unit, Unit> CheckUpdateClashMetaCoreCmd { get; }
+
+        //public ReactiveCommand<Unit, Unit> CheckUpdateClashCoreCmd { get; }
+        //public ReactiveCommand<Unit, Unit> CheckUpdateClashMetaCoreCmd { get; }
         public ReactiveCommand<Unit, Unit> CheckUpdateSingBoxCoreCmd { get; }
+
         public ReactiveCommand<Unit, Unit> CheckUpdateGeoCmd { get; }
 
         public ReactiveCommand<Unit, Unit> ReloadCmd { get; }
@@ -337,6 +340,18 @@ namespace v2rayN.ViewModels
             {
                 EditServer(true, EConfigType.Trojan);
             });
+            AddHysteria2ServerCmd = ReactiveCommand.Create(() =>
+            {
+                EditServer(true, EConfigType.Hysteria2);
+            });
+            AddTuicServerCmd = ReactiveCommand.Create(() =>
+            {
+                EditServer(true, EConfigType.Tuic);
+            });
+            AddWireguardServerCmd = ReactiveCommand.Create(() =>
+            {
+                EditServer(true, EConfigType.Wireguard);
+            });
             AddCustomServerCmd = ReactiveCommand.Create(() =>
             {
                 EditServer(true, EConfigType.Custom);
@@ -496,26 +511,26 @@ namespace v2rayN.ViewModels
             {
                 CheckUpdateN();
             });
-            CheckUpdateV2flyCoreCmd = ReactiveCommand.Create(() =>
-            {
-                CheckUpdateCore(ECoreType.v2fly_v5);
-            });
-            CheckUpdateSagerNetCoreCmd = ReactiveCommand.Create(() =>
-            {
-                CheckUpdateCore(ECoreType.SagerNet);
-            });
+            //CheckUpdateV2flyCoreCmd = ReactiveCommand.Create(() =>
+            //{
+            //    CheckUpdateCore(ECoreType.v2fly_v5);
+            //});
+            //CheckUpdateSagerNetCoreCmd = ReactiveCommand.Create(() =>
+            //{
+            //    CheckUpdateCore(ECoreType.SagerNet);
+            //});
             CheckUpdateXrayCoreCmd = ReactiveCommand.Create(() =>
             {
                 CheckUpdateCore(ECoreType.Xray);
             });
-            CheckUpdateClashCoreCmd = ReactiveCommand.Create(() =>
-            {
-                CheckUpdateCore(ECoreType.clash);
-            });
-            CheckUpdateClashMetaCoreCmd = ReactiveCommand.Create(() =>
-            {
-                CheckUpdateCore(ECoreType.clash_meta);
-            });
+            //CheckUpdateClashCoreCmd = ReactiveCommand.Create(() =>
+            //{
+            //    CheckUpdateCore(ECoreType.clash);
+            //});
+            //CheckUpdateClashMetaCoreCmd = ReactiveCommand.Create(() =>
+            //{
+            //    CheckUpdateCore(ECoreType.clash_meta);
+            //});
             CheckUpdateSingBoxCoreCmd = ReactiveCommand.Create(() =>
             {
                 CheckUpdateCore(ECoreType.sing_box);
@@ -558,7 +573,7 @@ namespace v2rayN.ViewModels
 
         private void Init()
         {
-            ConfigHandler.InitBuiltinRouting(ref _config);
+            ConfigHandler.InitBuiltinRouting(_config);
             ConfigHandler.InitBuiltinDNS(_config);
             _coreHandler = new CoreHandler(_config, UpdateHandler);
             Locator.CurrentMutable.RegisterLazySingleton(() => _coreHandler, typeof(CoreHandler));
@@ -621,8 +636,8 @@ namespace v2rayN.ViewModels
                         return;
                     }
 
-                    SpeedProxyDisplay = string.Format(ResUI.SpeedDisplayText, Global.agentTag, Utils.HumanFy(update.proxyUp), Utils.HumanFy(update.proxyDown));
-                    SpeedDirectDisplay = string.Format(ResUI.SpeedDisplayText, Global.directTag, Utils.HumanFy(update.directUp), Utils.HumanFy(update.directDown));
+                    SpeedProxyDisplay = string.Format(ResUI.SpeedDisplayText, Global.ProxyTag, Utils.HumanFy(update.proxyUp), Utils.HumanFy(update.proxyDown));
+                    SpeedDirectDisplay = string.Format(ResUI.SpeedDisplayText, Global.DirectTag, Utils.HumanFy(update.directUp), Utils.HumanFy(update.directDown));
 
                     if (update.proxyUp + update.proxyDown > 0)
                     {
@@ -639,13 +654,13 @@ namespace v2rayN.ViewModels
 
                                 if (SelectedProfile?.indexId == item.indexId)
                                 {
-                                    var temp = Utils.DeepCopy(item);
+                                    var temp = JsonUtils.DeepCopy(item);
                                     _profileItems.Replace(item, temp);
                                     SelectedProfile = temp;
                                 }
                                 else
                                 {
-                                    _profileItems.Replace(item, Utils.DeepCopy(item));
+                                    _profileItems.Replace(item, JsonUtils.DeepCopy(item));
                                 }
                             }
                         }
@@ -654,7 +669,7 @@ namespace v2rayN.ViewModels
             }
             catch (Exception ex)
             {
-                Utils.SaveLog(ex.Message, ex);
+                Logging.SaveLog(ex.Message, ex);
             }
         }
 
@@ -687,7 +702,7 @@ namespace v2rayN.ViewModels
                 {
                     item.speedVal = $"{speed} {Global.SpeedUnit}";
                 }
-                _profileItems.Replace(item, Utils.DeepCopy(item));
+                _profileItems.Replace(item, JsonUtils.DeepCopy(item));
             }
         }
 
@@ -721,10 +736,10 @@ namespace v2rayN.ViewModels
         {
             try
             {
-                Utils.SaveLog("MyAppExit Begin");
+                Logging.SaveLog("MyAppExit Begin");
 
                 StorageUI();
-                ConfigHandler.SaveConfig(ref _config);
+                ConfigHandler.SaveConfig(_config);
 
                 //HttpProxyHandle.CloseHttpAgent(config);
                 if (blWindowsShutDown)
@@ -742,7 +757,7 @@ namespace v2rayN.ViewModels
                 _statistics?.Close();
 
                 _coreHandler.CoreStop();
-                Utils.SaveLog("MyAppExit End");
+                Logging.SaveLog("MyAppExit End");
             }
             catch { }
             finally
@@ -822,7 +837,7 @@ namespace v2rayN.ViewModels
                             totalDown = t22 == null ? "" : Utils.HumanFy(t22.totalDown),
                             totalUp = t22 == null ? "" : Utils.HumanFy(t22.totalUp)
                         }).OrderBy(t => t.sort).ToList();
-            _lstProfile = Utils.FromJson<List<ProfileItem>>(Utils.ToJson(lstModel));
+            _lstProfile = JsonUtils.Deserialize<List<ProfileItem>>(JsonUtils.Serialize(lstModel));
 
             Application.Current.Dispatcher.Invoke((Action)(() =>
             {
@@ -844,7 +859,7 @@ namespace v2rayN.ViewModels
                 RefreshServersMenu();
 
                 //display running server
-                var running = ConfigHandler.GetDefaultServer(ref _config);
+                var running = ConfigHandler.GetDefaultServer(_config);
                 if (running != null)
                 {
                     var runningSummary = running.GetSummary();
@@ -928,7 +943,7 @@ namespace v2rayN.ViewModels
             }
             else
             {
-                lstSelecteds = Utils.FromJson<List<ProfileItem>>(Utils.ToJson(orderProfiles));
+                lstSelecteds = JsonUtils.Deserialize<List<ProfileItem>>(JsonUtils.Serialize(orderProfiles));
             }
 
             return 0;
@@ -982,7 +997,7 @@ namespace v2rayN.ViewModels
         public void AddServerViaClipboard()
         {
             var clipboardData = Utils.GetClipboardData();
-            int ret = ConfigHandler.AddBatchServers(ref _config, clipboardData!, _subId, false);
+            int ret = ConfigHandler.AddBatchServers(_config, clipboardData!, _subId, false);
             if (ret > 0)
             {
                 InitSubscriptionView();
@@ -1009,7 +1024,7 @@ namespace v2rayN.ViewModels
             }
             else
             {
-                int ret = ConfigHandler.AddBatchServers(ref _config, result, _subId, false);
+                int ret = ConfigHandler.AddBatchServers(_config, result, _subId, false);
                 if (ret > 0)
                 {
                     InitSubscriptionView();
@@ -1056,7 +1071,7 @@ namespace v2rayN.ViewModels
             {
                 return;
             }
-            if (ConfigHandler.CopyServer(ref _config, lstSelecteds) == 0)
+            if (ConfigHandler.CopyServer(_config, lstSelecteds) == 0)
             {
                 RefreshServers();
                 _noticeHandler?.Enqueue(ResUI.OperationSuccess);
@@ -1089,7 +1104,7 @@ namespace v2rayN.ViewModels
                 return;
             }
 
-            if (ConfigHandler.SetDefaultServerIndex(ref _config, indexId) == 0)
+            if (ConfigHandler.SetDefaultServerIndex(_config, indexId) == 0)
             {
                 RefreshServers();
                 Reload();
@@ -1121,7 +1136,7 @@ namespace v2rayN.ViewModels
                 _noticeHandler?.Enqueue(ResUI.PleaseSelectServer);
                 return;
             }
-            string url = ShareHandler.GetShareUrl(item);
+            var url = ShareHandler.GetShareUrl(item);
             if (Utils.IsNullOrEmpty(url))
             {
                 return;
@@ -1145,7 +1160,7 @@ namespace v2rayN.ViewModels
 
             _dicHeaderSort.TryAdd(colName, true);
             _dicHeaderSort.TryGetValue(colName, out bool asc);
-            if (ConfigHandler.SortServers(ref _config, _subId, colName, asc) != 0)
+            if (ConfigHandler.SortServers(_config, _subId, colName, asc) != 0)
             {
                 return;
             }
@@ -1155,7 +1170,7 @@ namespace v2rayN.ViewModels
 
         public void TestServerAvailability()
         {
-            var item = ConfigHandler.GetDefaultServer(ref _config);
+            var item = ConfigHandler.GetDefaultServer(_config);
             if (item == null || item.configType == EConfigType.Custom)
             {
                 return;
@@ -1209,7 +1224,7 @@ namespace v2rayN.ViewModels
             {
                 return;
             }
-            if (ConfigHandler.MoveServer(ref _config, ref _lstProfile, index, eMove) == 0)
+            if (ConfigHandler.MoveServer(_config, ref _lstProfile, index, eMove) == 0)
             {
                 RefreshServers();
             }
@@ -1220,7 +1235,7 @@ namespace v2rayN.ViewModels
             var targetIndex = _profileItems.IndexOf(targetItem);
             if (startIndex >= 0 && targetIndex >= 0 && startIndex != targetIndex)
             {
-                if (ConfigHandler.MoveServer(ref _config, ref _lstProfile, startIndex, EMove.Position, targetIndex) == 0)
+                if (ConfigHandler.MoveServer(_config, ref _lstProfile, startIndex, EMove.Position, targetIndex) == 0)
                 {
                     RefreshServers();
                 }
@@ -1350,7 +1365,7 @@ namespace v2rayN.ViewModels
             var ret = (new RoutingSettingWindow()).ShowDialog();
             if (ret == true)
             {
-                ConfigHandler.InitBuiltinRouting(ref _config);
+                ConfigHandler.InitBuiltinRouting(_config);
                 RefreshRoutingsMenu();
                 //RefreshServers();
                 Reload();
@@ -1386,22 +1401,17 @@ namespace v2rayN.ViewModels
 
         private void ImportOldGuiConfig()
         {
-            OpenFileDialog fileDialog = new()
-            {
-                Multiselect = false,
-                Filter = "guiNConfig|*.json|All|*.*"
-            };
-            if (fileDialog.ShowDialog() != true)
+            if (UI.OpenFileDialog(out string fileName,
+                "guiNConfig|*.json|All|*.*") != true)
             {
                 return;
             }
-            string fileName = fileDialog.FileName;
             if (Utils.IsNullOrEmpty(fileName))
             {
                 return;
             }
 
-            var ret = ConfigHandler.ImportOldGuiConfig(ref _config, fileName);
+            var ret = ConfigHandler.ImportOldGuiConfig(_config, fileName);
             if (ret == 0)
             {
                 RefreshRoutingsMenu();
@@ -1443,7 +1453,7 @@ namespace v2rayN.ViewModels
                     CloseV2ray();
 
                     string fileName = Utils.GetTempPath(Utils.GetDownloadFileName(msg));
-                    string toPath = Utils.GetBinPath("", type);
+                    string toPath = Utils.GetBinPath("", type.ToString());
 
                     FileManager.ZipExtractToFile(fileName, toPath, _config.guiItem.ignoreGeoUpdateCore ? "geo" : "");
 
@@ -1487,7 +1497,7 @@ namespace v2rayN.ViewModels
             {
                 _coreHandler.LoadCore();
 
-                //ConfigHandler.SaveConfig(ref _config, false);
+                //ConfigHandler.SaveConfig(_config, false);
 
                 ChangeSystemProxyStatus(_config.sysProxyType, false);
             });
@@ -1502,7 +1512,7 @@ namespace v2rayN.ViewModels
 
         private void CloseV2ray()
         {
-            ConfigHandler.SaveConfig(ref _config, false);
+            ConfigHandler.SaveConfig(_config, false);
 
             ChangeSystemProxyStatus(ESysProxyType.ForcedClear, false);
 
@@ -1523,7 +1533,7 @@ namespace v2rayN.ViewModels
             ChangeSystemProxyStatus(type, true);
 
             SystemProxySelected = (int)_config.sysProxyType;
-            ConfigHandler.SaveConfig(ref _config, false);
+            ConfigHandler.SaveConfig(_config, false);
         }
 
         private void ChangeSystemProxyStatus(ESysProxyType type, bool blChange)
@@ -1591,7 +1601,7 @@ namespace v2rayN.ViewModels
                 return;
             }
 
-            if (ConfigHandler.SetDefaultRouting(ref _config, item) == 0)
+            if (ConfigHandler.SetDefaultRouting(_config, item) == 0)
             {
                 _noticeHandler?.SendMessage(ResUI.TipChangeRouting, true);
                 Reload();
@@ -1616,6 +1626,13 @@ namespace v2rayN.ViewModels
             if (_config.tunModeItem.enableTun != EnableTun)
             {
                 _config.tunModeItem.enableTun = EnableTun;
+                // When running as a non-administrator, reboot to administrator mode
+                if (EnableTun && !Utils.IsAdministrator())
+                {
+                    _config.tunModeItem.enableTun = false;
+                    RebootAsAdmin();
+                    return;
+                }
                 Reload();
             }
         }
@@ -1697,7 +1714,7 @@ namespace v2rayN.ViewModels
                           {
                               _config.uiItem.colorModeDark = ColorModeDark;
                               ModifyTheme(ColorModeDark);
-                              ConfigHandler.SaveConfig(ref _config);
+                              ConfigHandler.SaveConfig(_config);
                           }
                       });
 
@@ -1708,10 +1725,14 @@ namespace v2rayN.ViewModels
                         if (_config.uiItem.followSystemTheme != FollowSystemTheme)
                         {
                             _config.uiItem.followSystemTheme = FollowSystemTheme;
-                            ConfigHandler.SaveConfig(ref _config);
+                            ConfigHandler.SaveConfig(_config);
                             if (FollowSystemTheme)
                             {
                                 ModifyTheme(!Utils.IsLightTheme());
+                            }
+                            else
+                            {
+                                ModifyTheme(ColorModeDark);
                             }
                         }
                     });
@@ -1732,7 +1753,7 @@ namespace v2rayN.ViewModels
                      {
                          _config.uiItem.colorPrimaryName = SelectedSwatch?.Name;
                          ChangePrimaryColor(SelectedSwatch.ExemplarHue.Color);
-                         ConfigHandler.SaveConfig(ref _config);
+                         ConfigHandler.SaveConfig(_config);
                      }
                  });
 
@@ -1750,7 +1771,7 @@ namespace v2rayN.ViewModels
                           Application.Current.Resources["StdFontSize2"] = size + 2;
                           Application.Current.Resources["StdFontSizeMsg"] = size - 1;
 
-                          ConfigHandler.SaveConfig(ref _config);
+                          ConfigHandler.SaveConfig(_config);
                       }
                   });
 
@@ -1763,7 +1784,7 @@ namespace v2rayN.ViewModels
                     {
                         _config.uiItem.currentLanguage = CurrentLanguage;
                         Thread.CurrentThread.CurrentUICulture = new(CurrentLanguage);
-                        ConfigHandler.SaveConfig(ref _config);
+                        ConfigHandler.SaveConfig(_config);
                     }
                 });
         }
@@ -1830,7 +1851,7 @@ namespace v2rayN.ViewModels
             if (_config.uiItem.autoHideStartup)
             {
                 Observable.Range(1, 1)
-                 .Delay(TimeSpan.FromSeconds(2))
+                 .Delay(TimeSpan.FromSeconds(0.5))
                  .Subscribe(x =>
                  {
                      Application.Current.Dispatcher.Invoke(() =>
