@@ -3,7 +3,6 @@ using ReactiveUI.Fody.Helpers;
 using Splat;
 using System.Reactive;
 using System.Windows;
-using v2rayN.Base;
 using v2rayN.Handler;
 using v2rayN.Mode;
 using v2rayN.Resx;
@@ -45,7 +44,7 @@ namespace v2rayN.ViewModels
             if (rulesItem.id.IsNullOrEmpty())
             {
                 rulesItem.id = Utils.GetGUID(false);
-                rulesItem.outboundTag = Global.agentTag;
+                rulesItem.outboundTag = Global.ProxyTag;
                 rulesItem.enabled = true;
                 SelectedSource = rulesItem;
             }
