@@ -6,7 +6,6 @@ using ReactiveUI.Fody.Helpers;
 using Splat;
 using System.Reactive;
 using System.Windows;
-using v2rayN.Base;
 using v2rayN.Handler;
 using v2rayN.Mode;
 using v2rayN.Resx;
@@ -104,7 +103,7 @@ namespace v2rayN.ViewModels
 
             foreach (var it in SelectedSources)
             {
-                ConfigHandler.DeleteSubItem(ref _config, it?.id);
+                ConfigHandler.DeleteSubItem(_config, it.id);
             }
             RefreshSubItems();
             _noticeHandler?.Enqueue(ResUI.OperationSuccess);
