@@ -2,7 +2,6 @@
 using System.Reactive.Disposables;
 using System.Windows;
 using System.Windows.Controls;
-using v2rayN.Base;
 using v2rayN.Handler;
 using v2rayN.Mode;
 using v2rayN.Resx;
@@ -35,14 +34,14 @@ namespace v2rayN.Views
 
             if (profileItem.configType == EConfigType.VLESS)
             {
-                Global.coreTypes4VLESS.ForEach(it =>
+                Global.CoreTypes4VLESS.ForEach(it =>
                 {
                     cmbCoreType.Items.Add(it);
                 });
             }
             else
             {
-                Global.coreTypes.ForEach(it =>
+                Global.CoreTypes.ForEach(it =>
                 {
                     cmbCoreType.Items.Add(it);
                 });
@@ -52,20 +51,20 @@ namespace v2rayN.Views
             cmbStreamSecurity.Items.Add(string.Empty);
             cmbStreamSecurity.Items.Add(Global.StreamSecurity);
 
-            Global.networks.ForEach(it =>
+            Global.Networks.ForEach(it =>
             {
                 cmbNetwork.Items.Add(it);
             });
-            Global.fingerprints.ForEach(it =>
+            Global.Fingerprints.ForEach(it =>
             {
                 cmbFingerprint.Items.Add(it);
                 cmbFingerprint2.Items.Add(it);
             });
-            Global.allowInsecures.ForEach(it =>
+            Global.AllowInsecures.ForEach(it =>
             {
                 cmbAllowInsecure.Items.Add(it);
             });
-            Global.alpns.ForEach(it =>
+            Global.Alpns.ForEach(it =>
             {
                 cmbAlpn.Items.Add(it);
             });
@@ -74,7 +73,7 @@ namespace v2rayN.Views
             {
                 case EConfigType.VMess:
                     gridVMess.Visibility = Visibility.Visible;
-                    Global.vmessSecuritys.ForEach(it =>
+                    Global.VmessSecuritys.ForEach(it =>
                     {
                         cmbSecurity.Items.Add(it);
                     });
@@ -99,7 +98,7 @@ namespace v2rayN.Views
                 case EConfigType.VLESS:
                     gridVLESS.Visibility = Visibility.Visible;
                     cmbStreamSecurity.Items.Add(Global.StreamSecurityReality);
-                    Global.flows.ForEach(it =>
+                    Global.Flows.ForEach(it =>
                     {
                         cmbFlow5.Items.Add(it);
                     });
@@ -112,10 +111,43 @@ namespace v2rayN.Views
                 case EConfigType.Trojan:
                     gridTrojan.Visibility = Visibility.Visible;
                     cmbStreamSecurity.Items.Add(Global.StreamSecurityReality);
-                    Global.flows.ForEach(it =>
+                    Global.Flows.ForEach(it =>
                     {
                         cmbFlow6.Items.Add(it);
                     });
+                    break;
+
+                case EConfigType.Hysteria2:
+                    gridHysteria2.Visibility = Visibility.Visible;
+                    sepa2.Visibility = Visibility.Collapsed;
+                    gridTransport.Visibility = Visibility.Collapsed;
+                    cmbCoreType.IsEnabled = false;
+                    cmbFingerprint.IsEnabled = false;
+                    cmbFingerprint.Text = string.Empty;
+                    break;
+
+                case EConfigType.Tuic:
+                    gridTuic.Visibility = Visibility.Visible;
+                    sepa2.Visibility = Visibility.Collapsed;
+                    gridTransport.Visibility = Visibility.Collapsed;
+                    cmbCoreType.IsEnabled = false;
+                    cmbFingerprint.IsEnabled = false;
+                    cmbFingerprint.Text = string.Empty;
+
+                    Global.TuicCongestionControls.ForEach(it =>
+                    {
+                        cmbHeaderType8.Items.Add(it);
+                    });
+                    break;
+
+                case EConfigType.Wireguard:
+                    gridWireguard.Visibility = Visibility.Visible;
+
+                    sepa2.Visibility = Visibility.Collapsed;
+                    gridTransport.Visibility = Visibility.Collapsed;
+                    gridTls.Visibility = Visibility.Collapsed;
+                    cmbCoreType.IsEnabled = false;
+
                     break;
             }
 
@@ -155,6 +187,25 @@ namespace v2rayN.Views
                     case EConfigType.Trojan:
                         this.Bind(ViewModel, vm => vm.SelectedSource.id, v => v.txtId6.Text).DisposeWith(disposables);
                         this.Bind(ViewModel, vm => vm.SelectedSource.flow, v => v.cmbFlow6.Text).DisposeWith(disposables);
+                        break;
+
+                    case EConfigType.Hysteria2:
+                        this.Bind(ViewModel, vm => vm.SelectedSource.id, v => v.txtId7.Text).DisposeWith(disposables);
+                        this.Bind(ViewModel, vm => vm.SelectedSource.path, v => v.txtPath7.Text).DisposeWith(disposables);
+                        break;
+
+                    case EConfigType.Tuic:
+                        this.Bind(ViewModel, vm => vm.SelectedSource.id, v => v.txtId8.Text).DisposeWith(disposables);
+                        this.Bind(ViewModel, vm => vm.SelectedSource.security, v => v.txtSecurity8.Text).DisposeWith(disposables);
+                        this.Bind(ViewModel, vm => vm.SelectedSource.headerType, v => v.cmbHeaderType8.Text).DisposeWith(disposables);
+                        break;
+
+                    case EConfigType.Wireguard:
+                        this.Bind(ViewModel, vm => vm.SelectedSource.id, v => v.txtId9.Text).DisposeWith(disposables);
+                        this.Bind(ViewModel, vm => vm.SelectedSource.publicKey, v => v.txtPublicKey9.Text).DisposeWith(disposables);
+                        this.Bind(ViewModel, vm => vm.SelectedSource.path, v => v.txtPath9.Text).DisposeWith(disposables);
+                        this.Bind(ViewModel, vm => vm.SelectedSource.requestHost, v => v.txtRequestHost9.Text).DisposeWith(disposables);
+                        this.Bind(ViewModel, vm => vm.SelectedSource.shortId, v => v.txtShortId9.Text).DisposeWith(disposables);
                         break;
                 }
                 this.Bind(ViewModel, vm => vm.SelectedSource.network, v => v.cmbNetwork.Text).DisposeWith(disposables);
@@ -236,7 +287,7 @@ namespace v2rayN.Views
             else if (network is "kcp" or "quic")
             {
                 cmbHeaderType.Items.Add(Global.None);
-                Global.kcpHeaderTypes.ForEach(it =>
+                Global.KcpHeaderTypes.ForEach(it =>
                 {
                     cmbHeaderType.Items.Add(it);
                 });
@@ -299,11 +350,6 @@ namespace v2rayN.Views
                     labHeaderType.Visibility = Visibility.Hidden;
                     break;
             }
-        }
-
-        private void btnCancel_Click(object sender, RoutedEventArgs e)
-        {
-            this.Close();
         }
     }
 }
