@@ -30,13 +30,14 @@ namespace v2rayN.Views
 
             ViewModel = new DNSSettingViewModel(this);
 
-            Global.domainStrategy4Freedoms.ForEach(it =>
+            Global.DomainStrategy4Freedoms.ForEach(it =>
             {
                 cmbdomainStrategy4Freedom.Items.Add(it);
             });
 
             this.WhenActivated(disposables =>
             {
+                this.Bind(ViewModel, vm => vm.useSystemHosts, v => v.togUseSystemHosts.IsChecked).DisposeWith(disposables);
                 this.Bind(ViewModel, vm => vm.domainStrategy4Freedom, v => v.cmbdomainStrategy4Freedom.Text).DisposeWith(disposables);
                 this.Bind(ViewModel, vm => vm.normalDNS, v => v.txtnormalDNS.Text).DisposeWith(disposables);
                 this.Bind(ViewModel, vm => vm.normalDNS2, v => v.txtnormalDNS2.Text).DisposeWith(disposables);
@@ -56,11 +57,6 @@ namespace v2rayN.Views
         private void linkDnsSingboxObjectDoc_Click(object sender, RoutedEventArgs e)
         {
             Utils.ProcessStart("https://sing-box.sagernet.org/zh/configuration/dns/");
-        }
-
-        private void btnCancel_Click(object sender, RoutedEventArgs e)
-        {
-            this.Close();
         }
     }
 }
