@@ -1,7 +1,6 @@
 ﻿using ReactiveUI;
 using System.Reactive.Disposables;
 using System.Windows;
-using v2rayN.Base;
 using v2rayN.Mode;
 using v2rayN.ViewModels;
 
@@ -29,10 +28,10 @@ namespace v2rayN.Views
             clbInboundTag.SelectionChanged += ClbInboundTag_SelectionChanged;
 
             ViewModel = new RoutingRuleDetailsViewModel(rulesItem, this);
-            cmbOutboundTag.Items.Add(Global.agentTag);
-            cmbOutboundTag.Items.Add(Global.directTag);
-            cmbOutboundTag.Items.Add(Global.blockTag);
-            Global.Protocols.ForEach(it =>
+            cmbOutboundTag.Items.Add(Global.ProxyTag);
+            cmbOutboundTag.Items.Add(Global.DirectTag);
+            cmbOutboundTag.Items.Add(Global.BlockTag);
+            Global.RuleProtocols.ForEach(it =>
             {
                 clbProtocol.Items.Add(it);
             });
@@ -80,11 +79,6 @@ namespace v2rayN.Views
         private void ClbInboundTag_SelectionChanged(object sender, System.Windows.Controls.SelectionChangedEventArgs e)
         {
             ViewModel.InboundTagItems = clbInboundTag.SelectedItems.Cast<string>().ToList();
-        }
-
-        private void btnCancel_Click(object sender, RoutedEventArgs e)
-        {
-            this.Close();
         }
 
         private void linkRuleobjectDoc_Click(object sender, RoutedEventArgs e)
